% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Estimate density & abundance}
\usage{
abundance(
  segments,
  sightings,
  das,
  strata,
  truncation_distance,
  ss_correction = 1,
  use_g0 = TRUE,
  g0 = c(1, 1),
  g0_threshold = 20,
  region_pool = FALSE,
  region_pool_name = "Regions pooled",
  region_pool_area = NA,
  year_pool = FALSE,
  forced_effort = NULL,
  verbose = TRUE
)
}
\arguments{
\item{segments}{Effort segments dataframe, drawn from a \code{cruz} object (e.g., \code{cruz$cohorts[[1]]$segments}),
already filtered to contain the effort you wish to use to estimate density/abundance.}

\item{sightings}{Sightings dataframe, drawn from a \code{cruz} object (e.g., \code{cruz$cohorts[[1]]$sightings}),
already filtered to contain the sightings you wish to use to fit the density/abundance.
This sightings dataframe must have a column named \code{esw}, which can be provided
by the \code{LTabundR} function \code{fit_df()} (fit a detection function model).
If \code{NA}'s occur in the \code{esw} column, they will be replaced with the mean \code{esw} value
for the remainder of the dataset in that region-year.
Similarly, if \code{sightings} has a column named \code{ss_valid} (all standard \code{cruz} objects do)
and any of the rows in that column are \code{FALSE}, those rows will have their \code{best}
school size estimate (which will be \code{NA} or \code{1}, since they are invalid) replaced
by the mean best estimate for their respective species.}

\item{das}{Dataframe of DAS data, drawn from a \code{cruz} object (e.g., \code{cruz$cohorts[[1]]$das})}

\item{strata}{A dataframe, drawn from a \code{cruz} object (e.g., \code{cruz$strata}),
summarizing the geostrata provided (their name and area, in square km).}

\item{truncation_distance}{The truncation distance to apply to sightings.}

\item{ss_correction}{Should a correction be applied to school sizes?
School sizes will be scaled by this number. The default, \code{1}, means no changes will occur.}

\item{use_g0}{A Boolean, with default \code{TRUE}, indicating whether or not to use custom \code{g(0)} value(s).
If \code{FALSE}, the assumed \code{g(0)} value will be 1.}

\item{g0}{A numeric vector of length 2: the \code{g(0)} for small and large groups.}

\item{g0_threshold}{The school size threshold between small and large groups.}

\item{region_pool}{Accepts a Boolean; if \code{TRUE}, the functions will produce a single estimate for all geostrata combined;
if \code{FALSE}, the function calculates density and abundance for each geostratum
contained within the data (according to \code{segments$stratum}) separately. If you set
\code{region_pool} to \code{TRUE}, the next two arguments ought to be specified.}

\item{region_pool_name}{A character string indicating a custom name to use in the \code{Region} column in the output.
Only used in the event that \code{region_pool == TRUE}.}

\item{region_pool_area}{A numeric indicating the area, in square km, of the pooled region.
Only used in the event that \code{region_pool == TRUE}.
If left \code{NULL} but \code{region_pool == TRUE}, abundance will not be calculated.}

\item{year_pool}{A Boolean indicating whether or not to pool years (set \code{year_pool} to \code{TRUE})
or to report density/abundance for each year separately (set to \code{FALSE}).}

\item{forced_effort}{If this is a single numeric value instead of \code{NULL} (\code{NULL} is the default),
this value will be used as the survey effort, in km, in a brute-force method;
this same value will be used for every year and region. This is only helpful if you
are looking for a relatively easy way to compare results from your own analysis to another
(e.g., comparing \code{LTabundR} results to reports from NOAA reports prior to 2021, in which effort was
calculated slightly differently).}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A \code{data.frame} with a row for each density/abundance estimate.
You may expect multiple rows if multiple regions or years are contained within the datasets provided.
This \code{data.frame} contains the following fields:
\enumerate{
\item \code{Region}: Name(s) of geostrata represented in this estimate.
\item \code{Area}: Area of geostratum / region, in square km.
\item \code{year}: Years represented in this estimate.
\item \code{segments}: The number of effort segments used to estimate density/abundance.
\item \code{km}: The km of trackline effort contained in these segments.
\item \code{Area_covered}: The Area surveyed, according to \code{km} and \code{ESW_mean} (see next column).
\item \code{ESW_mean}: Mean effective strip width, in kw, calculated as the truncation distance multiplied by the mean probability of detection for all detections.
\item \code{n}: The number of detections in the data.
\item \code{g0_mean}: The mean \code{g(0)} estimate.
\item \code{ER_clusters}: The encounter rate for detections (schools) (\code{n / km})
\item \code{D_clusters}: The density of detections (schools).
\item \code{N_clusters}: The abundance of schools.
\item \code{size_mean}: Average school size.
\item \code{size_sd}: Standard deviation of school size.
\item \code{ER}: Animal encounter rate.
\item \code{D}: Animal density.
\item \code{N}: Animal abundance.
}
}
\description{
This function is typically not called by the user (though it certainly can be);
instead, it is called as a subroutine within \code{lta()}.
This function estimates density/abundance for \code{Wincruz} data,
as processed by \code{LTabundR::process_surveys()}.
}
