% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barlow_2015}
\alias{barlow_2015}
\title{Relative trackline detection probabilities from Barlow (2015).}
\format{
An object of class \code{data.frame} with 161 rows and 9 columns.
}
\usage{
barlow_2015
}
\description{
Relative trackline detection probabilities from Barlow (2015).
}
\details{
This dataset holds the results from Barlow (2015), "Inferring trackline detection
probabilities, g(0), for cetaceans from apparent densities in different survey
conditions" (\emph{Marine Mammal Science}), which used NOAA/NMFS cruise data from
1986 to 2010.
}
\keyword{datasets}
