% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{bca}
\alias{bca}
\title{BCA method function}
\usage{
bca(theta, conf.level = 0.95)
}
\arguments{
\item{theta}{Expectation.}

\item{conf.level}{Confidence level.}
}
\value{
A two-element vector with the lower and upper confidence interval calculated using the BCA method.
}
\description{
Replicated from the \code{coxed} package, which has been archived on \code{CRAN}.
Moving code here so that \code{LTabundR} can still use this function.
}
