% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnp_150km_1986_2020}
\alias{cnp_150km_1986_2020}
\title{Cruise data processed for the CNP, 1986 - 2020 (150 km segment lengths)}
\format{
An object of class \code{list} of length 3.
}
\usage{
cnp_150km_1986_2020
}
\description{
Cruise data processed for the CNP, 1986 - 2020 (150 km segment lengths)
}
\details{
This dataset was processed with the following code:\preformatted{data(group_size_coefficients)

survey <- load_survey_settings(
out_handling = 'remove',
max_row_interval = Inf,
segment_method = "equallength",
segment_target_km = 150,
segment_max_interval = 24,
segment_remainder_handling = c("segment"),
ship_list = NULL, # use package list
species_codes = NULL, # use package codes
group_size_coefficients = group_size_coefficients, # use package coefficients
smear_angles = FALSE)

data(strata_cnp)
strata <- strata_cnp

all_species <- load_cohort_settings(
id = "all",
species = NULL, #spp_codes,
strata = c('MHI', 'WHICEAS', 'HI_EEZ', 'OtherCNP'),
probable_species = FALSE,
sighting_method = 0,
cue_range = 0:7,
school_size_range = c(0, 10000),
school_size_calibrate = TRUE,
calibration_floor = 0,
use_low_if_na = TRUE,
io_sightings = 0,
geometric_mean_group = TRUE,
truncation_km = 7.5,
beaufort_range = 0:6,
abeam_sightings = TRUE,
strata_overlap_handling = c("smallest"),
distance_types = c('S','F','N'),
distance_modes = c('P','C'),
distance_on_off = TRUE
)

bottlenose <- load_cohort_settings(
id = "bottlenose",
species = '018',
strata = c('MHI', 'WHICEAS', 'HI_EEZ', 'OtherCNP', 'Bottlenose_BI', 'Bottlenose_OUFI', 'Bottlenose_KaNi'),
probable_species = FALSE,
sighting_method = 0,
cue_range = 0:7,
school_size_range = c(0, 10000),
school_size_calibrate = TRUE,
calibration_floor = 0,
use_low_if_na = TRUE,
io_sightings = 0,
geometric_mean_group = TRUE,
truncation_km = 7.5,
beaufort_range = 0:6,
abeam_sightings = TRUE,
strata_overlap_handling = c("smallest"),
distance_types = c('S','F','N'),
distance_modes = c('P','C'),
distance_on_off = TRUE
)

spotted <- load_cohort_settings(
id = "spotted",
species = '002',
strata = c('MHI', 'WHICEAS', 'HI_EEZ', 'OtherCNP','Spotted_OU','Spotted_FI','Spotted_BI'),
probable_species = FALSE,
sighting_method = 0,
cue_range = 0:7,
school_size_range = c(0, 10000),
school_size_calibrate = TRUE,
calibration_floor = 0,
use_low_if_na = TRUE,
io_sightings = 0,
geometric_mean_group = TRUE,
truncation_km = 7.5,
beaufort_range = 0:6,
abeam_sightings = TRUE,
strata_overlap_handling = c("smallest"),
distance_types = c('S','F','N'),
distance_modes = c('P','C'),
distance_on_off = TRUE
)

fkw <- load_cohort_settings(
id = "pseudorca",
species = '033',
strata = c('MHI', 'WHICEAS', 'HI_EEZ', 'OtherCNP','NHWI'),
probable_species = FALSE,
sighting_method = 0,
cue_range = 0:7,
school_size_range = c(0, 10000),
school_size_calibrate = TRUE,
calibration_floor = 0,
use_low_if_na = TRUE,
io_sightings = 0,
geometric_mean_group = TRUE,
truncation_km = 7.5,
beaufort_range = 0:6,
abeam_sightings = TRUE,
strata_overlap_handling = c("smallest"),
distance_types = c('S','F','N'),
distance_modes = c('P','C'),
distance_on_off = TRUE
)

settings <- load_settings(strata = strata,
                          survey = survey,
                          cohorts = list(all_species,
                                         bottlenose,
                                         spotted,
                                         fkw))

das_file = '../test_code/CNP/CenPac1986-2020_Final_alb.das'

cnp_150km_1986_2020 <- process_surveys(das_file = das_file,
                     settings = settings)
}
}
\keyword{datasets}
