% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnp_1986_2020_edits}
\alias{cnp_1986_2020_edits}
\title{Example of coded edits to \code{cruz} object}
\format{
An object of class \code{list} of length 4.
}
\usage{
cnp_1986_2020_edits
}
\description{
Example of coded edits to \code{cruz} object
}
\details{
This is an example of the kind of object that can be passed to the \code{edits} input in \code{process_surveys()}.
}
\keyword{datasets}
