% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{Coastline}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 4133 rows and 4 columns.
}
\usage{
coastline
}
\description{
Coastline
}
\details{
World coastline, downloadeded from Natural Earth.
}
\keyword{datasets}
