% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coordinates_in_strata.R
\name{coordinates_in_strata}
\alias{coordinates_in_strata}
\title{Determine whether coordinates occur inside a stratum polygon}
\usage{
coordinates_in_strata(lon, lat, poli)
}
\arguments{
\item{lon}{Longitude coordinates in decimal degrees, in which West coordinates are negative.}

\item{lat}{Latitude coordinates in decimal degrees, in which South coordinates are negative.  Must be the same length as \code{lon}.}

\item{poli}{A dataframe of polygon coordinates.}
}
\value{
A logical vector, the same length as the \code{lon},
indicating whether or not the coordinate falls within the polygon.
If the coordinate falls exactly on the polygon boundary, it will be counted as in.
}
\description{
This is an internal function typically not called by a user directly.
}
