% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruz_explorer.R
\name{cruz_explorer}
\alias{cruz_explorer}
\title{Launch a data explorer app for WinCruz surveys}
\usage{
cruz_explorer(cruz, cohort = 1)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}
}
\value{
A \code{Shiny} app for interactive data exploration.
This app allows you to filter the \code{cruz} object according to various fields,
then explore summary tables of effort, sightings, and detection distances.
There are also tabs for reviewing the data in tabular form and exploring an interactive \code{leaflet} map
of survey tracks, geostrata, and sightings.
}
\description{
Launch a data explorer app for WinCruz surveys
}
