% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruz_structure.R
\name{cruz_structure}
\alias{cruz_structure}
\title{Review of \code{cruz} object structure}
\usage{
cruz_structure(cruz)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}
}
\value{
Nothing; messages are printed to the Console.
}
\description{
This function prints an overview of the list structure and sample sizes
within your \code{cruz} object, which is produced from \code{LTabundR::process_surveys()}.
}
