% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_editor.R
\name{das_editor}
\alias{das_editor}
\title{Apply edits to a \code{DAS} file...}
\usage{
das_editor(edits)
}
\arguments{
\item{edits}{A list of sublists, each being a staged edit. Each sublist requires the following named slots:
\itemize{
\item \code{das_file}: Filepath to a \code{DAS} file.
\item \code{type}: The type of edit. Currently accepted options are:
\itemize{
\item \code{"text"}: the \code{edit} will be interpreted verbatim as text that will replace the specified data.
Both \code{rows} and \code{characters} must be supplied, as well as the \code{edit} itself.
\item \code{"function"}: the \code{edit} will be evaluated as a function that is applied to each row of specified data.
Both \code{rows} and \code{characters} must be supplied, as well as \code{edit.}
For example, \code{"tolower"} will convert the specified data to all lower case.
Another example: \code{"function(x){LTabundR::das_time(x, -10)}"}
will subtract 10 hours from the data if your specified \code{chars} is the date-timestamp for a row.
\item \code{"move"}: the \code{rows} will be deleted from their current location and pasted
immediately below the row number specified by \code{edit}.
The moved rows will be given the same date, time, latitude, and longitude, as the \code{edit} row.
Only \code{rows} and \code{edits} need to be supplied.
\item \code{"copy"}: the \code{rows} will be copied from their current location and pasted
immediately below the row number specified by \code{edit}.
The paste rows will be given the same date, time, latitude, and longitude, as the \code{edit} row.
Only \code{rows} and \code{edits} need to be supplied.
\item \code{"insert"}: the text provided in \code{edit} will be inserted verbatim immediately below
the first of the \code{rows} provided. Only \code{rows} and \code{edits} needs to be supplied.
\item \code{"delete"}: the \code{rows} will be deleted. Only \code{rows} needs to be supplied.
}
\item \code{rows}: A numeric vector of the rows to be edited.
\item \code{chars}: The character indices to be edited; e.g., \code{"1:3"} specifies that only the first three characters in each of \code{rows} will be edited.
\item \code{edit}: The actual edit, provided as a vector whose format depends on the \code{type} chosen above. See examples.
}
This \code{edits} list of lists can be as long as necessary, and it can contain as edits for as many \code{DAS} files as needed.
The \code{das} table that is returned by the function will concatenate all \code{das} files represented by the edits into a single table.}
}
\value{
A list with two named slots:
\itemize{
\item \code{das}: A \code{data.frame} containing the modified \code{DAS} data (with one column: \verb{$das}).
\item \code{log}: The \code{edits} input, with each sublist augmented with a \code{results} slot that shows the results of the edit.
}
}
\description{
...without modifying the original data. Edits are provided as a list, which the
user can store as a \code{RData} object or simply prepare with an \code{R} script. This
function loops through each edit and applies them to the respective \code{DAS} file.
This allows survey data to be modified reproducibly before being processed
with \code{LTabundR::process_surveys()}, without touching the original \code{DAS} data files or
requiring analysts to duplicate files and make one-off modifications manually.
}
