% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_editor_tofile.R
\name{das_editor_tofile}
\alias{das_editor_tofile}
\title{Edit \code{DAS} files and save edited versions}
\usage{
das_editor_tofile(das_file, edits, suffix = "_edited", verbose = TRUE)
}
\arguments{
\item{das_file}{A vector of \code{DAS} data filename(s).}

\item{edits}{A list of sublists, each being instructions for a staged edit. See \code{?das_editor()} for details.}

\item{suffix}{The suffix to append to \code{DAS} files to distiguish the original file from the new edited file.}

\item{verbose}{Boolean. Printed updates to the Console?}
}
\value{
At the end of the process, the function returns a vector of \code{DAS} filenames that should be passed to processing functions such as \code{process_surveys()} or \code{das_load()}.
}
\description{
This function is basically a wrapper for the function \code{das_editor()}, but this function saves the edited DAS data to new files.
It takes a set of \code{DAS} files and a set of editing instructions (provided as a list -- see \code{das_editor()} documentation),
applies the edits to each \code{DAS} file, and saves new versions of the DAS file. It does \emph{NOT} replace the original DAS file.
}
