% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_format.R
\name{das_format}
\alias{das_format}
\title{Finalize formatting of a \code{DAS} object}
\usage{
das_format(cruz, verbose = FALSE)
}
\arguments{
\item{cruz}{The list produced by \code{process_strata()}, with two slots: \code{settings} and \code{das}.}

\item{verbose}{Boolean, with default \code{FALSE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A list with nascent \code{cruz} object structure, with these slots:
\enumerate{
\item \code{settings} holds your settings object;
\item \code{strata} holds a dataframe summarizing the name and area (square km) of each geo-stratum, if any are provided;
\item \code{cohorts} holds a list with slots for each cohort of species as specified in \code{settings}.
Each cohort slot has a copy of the \code{DAS} data with a new \code{stratum} column,
with a stratum assignment tailored to its cohort-specific settings (specifically, the setting \code{stratum_overlap_handling}).
This list structure will be expanded upon in subsequent steps of \code{process_surveys()}.
}
}
\description{
This function removes invalid rows of data (due to missing Cruise numbers, timestamps, location coordinates, etc.),
calculates the distance, in kilometers, between each row of data,
adds a \code{ship} column with the ship name associated with the cruise number,
and initiates the \code{cruz} object structure, with a new \code{cohorts} slot.
\cr \cr
This is an internal function typically not called by a user directly.
It is the third subroutine called within \code{process_surveys()}; after \code{process_strata()} and before \code{segmentize()}.
}
