% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_inspector.R
\name{das_inspector}
\alias{das_inspector}
\title{Use a \code{Shiny} app to explore a \code{DAS} file and build up a list of edits. Those edits
will be saved as a list of instructions, which can subsequently be provided as an argument to
\code{LTabundR::process_surveys()}, which will apply your list of edits before processing the data.
This framework allows you to record details of required edits reproducibly without ever
modifying the original dataset or producing modified versions of the original.}
\usage{
das_inspector(das_file)
}
\arguments{
\item{das_file}{Path to your \code{DAS} file.}
}
\value{
A \code{Shiny} app is launched, where the user explores and builds up a list of staged edits.
Once the user closes that app, this function returns that list of edits. We recommend saving that return
in an object for easy call back later. To understand the structure of the list of edits, see \code{LTabundR::das_editor()}.
}
\description{
Use a \code{Shiny} app to explore a \code{DAS} file and build up a list of edits. Those edits
will be saved as a list of instructions, which can subsequently be provided as an argument to
\code{LTabundR::process_surveys()}, which will apply your list of edits before processing the data.
This framework allows you to record details of required edits reproducibly without ever
modifying the original dataset or producing modified versions of the original.
}
