% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_readtext.R
\name{das_readtext}
\alias{das_readtext}
\title{Read DAS file without parsing columns}
\usage{
das_readtext(das_file)
}
\arguments{
\item{das_file}{Filepath to \code{DAS} file.}
}
\value{
A one-column \code{data.frame} (column name is \code{das}).
}
\description{
In contrast to \code{swfscDAS::das_read()}, this function reads the DAS text in, saving
each line of data to a single row of a one-column \code{data.frame} (no column parsing).
This is useful if you wish to edit and save an altered version of the data using
\code{LTabundR::das_inspector()} and \code{LTabundR::das_editor()}. This function relies upon
the \code{readtext} package.
}
