% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_time.R
\name{das_time}
\alias{das_time}
\title{Adjust the timestamp within a \code{DAS} file}
\usage{
das_time(das, tz_adjust = "from utc")
}
\arguments{
\item{das}{A row of \code{DAS} data. It may already be cropped to the characters
corresponding to timestamp, date, lat, and long (characters 6 - 40).}

\item{tz_adjust}{One of two options: (1) a single numeric value indicating
the hours to add to or subtract from the date-time in each \code{DAS} row, or
(2) the character string \code{"from utc"}, which will indicate that the \code{DAS} timestamps
are in UTC and need to be converted to local time. To do this, the latitude and longitude
of each row will be used to determine the timezone of the event, then adjust the
timestamp from UTC accordingly. This process can take several minutes.}
}
\value{
A two-slot list: \verb{[[1]] dt} returns the revised date-time characters.
\verb{[[2]] das} returns the revised row in full (if a full row was originally provided).
}
\description{
This function is typically called within \code{LTabundR::das_inspector()} as a way of
staging an edit that adjusts the timestamps of certain rows. Not typically called directly by a user.
}
