% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_fit.R
\name{df_fit}
\alias{df_fit}
\title{Fit a detection function to \code{WinCruz} data}
\usage{
df_fit(
  sightings,
  truncation_distance = Inf,
  covariates = NULL,
  detection_function_base = "hn",
  base_model = "~1",
  delta_aic = 2,
  toplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sightings}{Sightings dataframe, drawn from a \code{cruz} object (e.g., \code{cruz$cohorts[[1]]$sightings}),
already filtered to contain the sightings you wish to use to fit the detection function.}

\item{truncation_distance}{The truncation distance to apply to sightings before model fitting.}

\item{covariates}{Covariates you wish to include as candidates in detection function models,
provided as a character vector. The covariates must match columns existing within \code{sightings}.
Common covariates that you will find within \code{sightings} include \code{c('Bft','LnSsTot','Cruise','Year','Ship','species')}.
Note that the function will ignore case, coercing all covariates to lowercase.}

\item{detection_function_base}{The base key for the deteion function, provided as a character vector.
Accepted values are \code{"hn"} (half-normal key, the default, which exhibit greater stability
when fitting to cetacean survey data; Gerrogette and Forcada 2005), \code{"hr"} (hazard-rate),
or \verb{c("hn", "hr)}, which will loop through both keys and attempt model fitting.}

\item{base_model}{The initial model formula, upon which to build using candidate covariates,
provided as a character vector. The default is \code{"~ 1"}.}

\item{delta_aic}{The AIC difference between the model yielding the lowest AICc and
other candidate models, used to define the best-fitting models. Typically,
AICc differences of less than 2 indicate effectively equal model performance.}

\item{toplot}{Boolean, with default \code{TRUE}, indiciating whether detection function plots (\code{Distance::plot.ds()})
should be displayed as the candidate models are tested.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A named list:
\enumerate{
\item \code{best_models}: A \code{data.frame} summary of the best-fitting models,
based upon the table produced by \code{Distance::summarize_ds_models()}.
See that function's documentation for details.
\item \code{all_models}:  Similar to the preceding slot, a tabular summary of \emph{all} models tested.
\item \code{best_objects}:  A list containing the \code{ds} objects (produced by package \code{Distance})
for each of the best-fitting models.
\item \code{sample_size}: A \code{data.frame} with the detections for each species within the species pool
used to fit the detection function.
\code{Ntot} is total detections for each species; \code{Ndet} is total detections within the truncation distance
and therefore used in the detection function fitting routine; \code{TD} is the truncation distance.
\item \code{tables}: A list of the data tables passed to \code{Distance::ds()} during model fitting.
\item \code{sightings}: The \code{sightings} \code{data.frame} provided, now with an \code{esw} columns provided estimated Effective Strip half-Width for each sighting.
}
}
\description{
This function is not typically called by the user (though it can be);
instead it is called as a subroutine from within the function \code{lta()}.
This function is a wrapper for the detection function fitting routine in the package
\code{mrds} (function \code{mrds::ddf()}). As a wrapper, this function conducts
automated stepwise model-fitting, trying all candidate function keys and covariates,
then selecting the best model(s) based upon \code{AICc}.
}
\details{
Model fitting is done in a forward stepwise procedure, starting fresh with each base key
provided. In the first round, the base model (no covariates) is fit first.
In the second round, each covariate is added one at a time; the covariate, if any, that produces the lowest AICc
below the AIC from the previous round is added to the formula.
This process is repeated in subsequent rounds until the AICc no longer improves.
All models within \code{delta_aic} of the model with the lowest AICc qualify as best-fitting models.
}
