% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_plot.R
\name{df_plot}
\alias{df_plot}
\title{Plot the best-fit detection function model(s)}
\usage{
df_plot(
  lta_result,
  model_colors = "black",
  model_pch = 1,
  hist_bars = 15,
  hist_col = "grey90",
  hist_border = "grey60",
  pt_show = 1,
  pt_cex = 0.8,
  pt_alpha = 0.3,
  line_col = "darkblue",
  line_lty = 1,
  line_lwd = 2.5,
  line_alpha = 0.7,
  bootstrap_show = TRUE,
  bootstrap_col = "steelblue4",
  bootstrap_alpha = 0.2,
  bootstrap_lwd = 0.5,
  main_show = TRUE,
  main = NULL,
  main_cex = 1.25,
  legend_show = FALSE,
  ymax = NULL,
  legend_x = 3.2,
  legend_y = 1.2,
  legend_cex = 0.8,
  xlab = "Distance (km) from trackline"
)
}
\arguments{
\item{lta_result}{The result of \code{lta()} for a single species pool.}

\item{model_colors}{A vector of colors; if more than one is provided,
the vector must be the same length as the number of best-fitting models
(\code{lta_result$df$best_models})}

\item{model_pch}{A vector of point \code{pch}, of either length 1 or the same as the
number of best-fitting models.}

\item{hist_bars}{Number of histogram bars to show.}

\item{hist_col}{Color of histogram bars.}

\item{hist_border}{Border color of histogram bars.}

\item{pt_show}{If \code{0}, no detection probabilities for individual detections will be shown.
If \code{1}, only detection probabilities for the first best-fitting model will be shown, as points.
(This is the default.)
If \code{2}, points will be displayed for \emph{all} best-fitting models.}

\item{pt_cex}{The size of detection points, if shown.}

\item{pt_alpha}{The transparency (0 = transparent; 1=solid) of points, if shown.}

\item{line_col}{Color of detection function line.}

\item{line_lty}{The line type for the detection function, following conventional base plot \code{lty} values.}

\item{line_lwd}{Detection function line thickness.}

\item{line_alpha}{Detection function line transparency, as a fraction between 0 and 1.}

\item{bootstrap_show}{Show detection function curves for all bootstrap iterations? Default is \code{TRUE}.}

\item{bootstrap_col}{Color of bootstrap detection function curves, if shown.}

\item{bootstrap_alpha}{Transparency of bootstrap detection function curves, if shown.}

\item{bootstrap_lwd}{Thickness of bootstrap detection function curves, if shown.}

\item{main_show}{Show a main title? \code{TRUE} (default) or \code{FALSE}.}

\item{main}{The text for the main title; if \code{NULL}, the species pool name will be provided.}

\item{main_cex}{Size of main title.}

\item{legend_show}{Show a legend, indicating the line/point/color for each best-fitting model?
\code{TRUE} or \code{FALSE} (default).}

\item{ymax}{Maximum of y-axis (can be helpful when making space for the legend). Can be \code{NULL}.}

\item{legend_x}{Left-right position of legend, if shown.}

\item{legend_y}{Up-down position of legend, if shown.}

\item{legend_cex}{Size of legend text, if shown.}

\item{xlab}{Text for x axis label, if shown.}
}
\value{
A plot is printed.
}
\description{
This function adapts \code{mrds::plot.ds()} to plot the detection functions fit and
provided by the output of \code{lta()} and \code{lta_multistock()}. Arguments allow for stylizing
the plot as well as displaying multiple best-fitting models on the same histogram.
}
