% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eez}
\alias{eez}
\title{EEZ (all US EEZ boundaries)}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 260 rows and 17 columns.
}
\usage{
eez
}
\description{
EEZ (all US EEZ boundaries)
}
\details{
World dataset of geopolitical Exclusive Economic Zones (EEZ), downloaded from Natural Earth.
}
\keyword{datasets}
