% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eez_ccs}
\alias{eez_ccs}
\title{EEZ - California Current System (formatted for \code{sf} mapping)}
\format{
An object of class \code{sfc_MULTILINESTRING} (inherits from \code{sfc}) of length 17.
}
\usage{
eez_ccs
}
\description{
EEZ - California Current System (formatted for \code{sf} mapping)
}
\details{
The EEZ relevant to the California Current System only, formatted for \code{sf} compatibility.
}
\keyword{datasets}
