% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eez_hawaii}
\alias{eez_hawaii}
\title{EEZ - Hawaii (formatted for \code{sf} mapping)}
\format{
An object of class \code{sfc_MULTILINESTRING} (inherits from \code{sfc}) of length 25.
}
\usage{
eez_hawaii
}
\description{
EEZ - Hawaii (formatted for \code{sf} mapping)
}
\details{
The EEZ relevant to the Hawai'ian area only, formatted for \code{sf} compatibility.
}
\keyword{datasets}
