% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_simulator.R
\name{er_simulator}
\alias{er_simulator}
\title{Encounter Rate simulator test}
\usage{
er_simulator(
  spp,
  cohort = 1,
  cruz,
  iterations = 1000,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{spp}{Species code}

\item{cohort}{The cohort whose data you would like to analyze, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{iterations}{Number of iterations}

\item{seed}{Set a seed (any integer) to ensure that the result is reproducible.
If left \code{NULL}, the results are liable to differ for each run of this function.}

\item{verbose}{Boolean; print updates to the console?}
}
\value{
A \code{dataframe} with a row for each year. Columns provide the number of observations of
the species of interest during systematic effort, and the p-value of the test.
The p-value represents the fraction of simulated encounter rates that exceed the observed encounter rate.
}
\description{
Test for the probability that year-to-year changes observed in a species' encounter rate
are due to random sampling variation instead of an actual change in the encounter rate.
This function uses bootstrap sampling of survey segments to see if random variation in
sampling could possibly produce an \emph{apparent} but immaterial change in encounter rate across years.
}
\details{
See the Appendix to Bradford et al. (2020) for analytical details, but briefly:
in each bootstrap iteration, survey segments are resampled in a way that preserves
the proportion of effort occurring within each geostratum in the data.
The resampled data are used to calculate the overall encounter rate across all years,
since the null hypothesis is that the encounter rate does not change across years.
This overall encounter rate is used to predict the number of sightings in each year,
based on the distance covered by the resampled segments in each year.
This process is repeated (\code{iterations} times) to produce a distribution of predicted sighting counts
in each year. This distribution is compared to the actual number of sightings observed in each year.
The number of simulated sightings counts that exceed the observed count reflects
the probability that the observed count is due to random sample variation alone.
}
