% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cruz}
\alias{example_cruz}
\title{Example cruz object (WHICEAS 2020)}
\format{
An object of class \code{list} of length 3.
}
\usage{
example_cruz
}
\description{
Example cruz object (WHICEAS 2020)
}
\details{
This example \code{cruz} object was prepared with the following code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{data(example_settings)
example_cruz <- process_surveys('data-raw/data/HICEASwinter2020.das',
                          settings = example_settings)
}\if{html}{\out{</div>}}
}
\keyword{datasets}
