% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_settings}
\alias{example_settings}
\title{Example settings used in \code{LTabundR} vignette (WHICEAS 2020)}
\format{
An object of class \code{list} of length 3.
}
\usage{
example_settings
}
\description{
Example settings used in \code{LTabundR} vignette (WHICEAS 2020)
}
\details{
This example \code{settings} object was prepared with the following code:\preformatted{data(strata_cnp)
data(study_cnp)
data(group_size_coefficients)
survey <- load_survey_settings()
cohort1 <- load_cohort_settings(strata = c('OtherCNP', 'HI_EEZ', 'WHICEAS'))
example_settings <- load_settings(strata = strata_cnp,
                               study_area = study_cnp,
                               survey = survey,
                               cohorts=list(cohort1))
}
}
\keyword{datasets}
