% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cruz.R
\name{filter_cruz}
\alias{filter_cruz}
\title{Filter \code{cruz} objects by species/region/year, etc.}
\usage{
filter_cruz(
  cruz,
  analysis_only = FALSE,
  spp = NULL,
  years = NULL,
  not_years = NULL,
  cruises = NULL,
  not_cruises = NULL,
  regions = NULL,
  not_regions = NULL,
  bft_range = NULL,
  eff_types = NULL,
  on_off = NULL,
  lat_range = NULL,
  lon_range = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{analysis_only}{If \code{TRUE}, data will be filtered to those viable for use in formal analysis.
For \code{segments}, this means rows in which \code{segments$use == TRUE}.
For \code{sightings}, this means rows in which \code{sightings$included == TRUE}.}

\item{spp}{A character vector of species codes, used to filter \code{sightings} based on \code{sightings$species}.}

\item{years}{A numeric vector of years, used to filter both \code{segments} and \code{sightings} to include only data from these years.}

\item{not_years}{A numeric vector of years, used to filter both \code{segments} and \code{sightings} to include only data \emph{not} from these years.
Usually not useful and can be left as \code{NULL}, but may be useful if your data contain many years and it can be
more efficient to exclude certain years rather than specify most years in the preceding argument.}

\item{cruises}{A numeric vector of cruise numbers, used to filter a la \code{years} above.}

\item{not_cruises}{A numeric vector of cruise numbers, used to filter a la \code{not_years} above.}

\item{regions}{A character vector of geostratum names, used to filter both \code{segments} and \code{sightings}.
Any segment or sighting occurring within \emph{any} (\emph{not} all) of the provided \code{regions} will be returned.
This holds for nested regions; for example, in analyses from the Central North Pacific, in which the Hawaii EEZ geostratum (\code{"HI-EEZ"})
is nested within the larger geostratum representing the entire CNP study area (\code{"OtherCNP"}),
an input of \code{regions = "OtherCNP"} will return segments/sightings \emph{both} inside the Hawaii EEZ \emph{and} outside of it.}

\item{not_regions}{A character vector of geostratum names, similar to above.
Any segment or sighting occurring within any of these \code{not_regions} will not be returned.
Using the example above, if \code{regions = "OtherCNP"} and \code{not_regions = "HI-EEZ"},
only segments occuring within \code{OtherCNP} \emph{and} outside of \code{HI-EEZ} will be returned.
This can be particularly useful for abundance estimates for pelagic stock that exclude nested insular stocks.}

\item{bft_range}{The  Beaufort Sea State values to filter to, provided as a numeric vector.}

\item{eff_types}{The effort types ('S' = Systematic, 'F' = Fine-scale, 'N' = Non-systematic) to filter to, provided as a character vector.}

\item{on_off}{The \code{OnEffort} values to filter to (\code{TRUE} and/or \code{FALSE}), provided as a logical vector.}

\item{lat_range}{If not \code{NULL}, a two-element numeric vector indicating latitudinal range of acceptable data. Accepted values are -90 (S) to 90 (N).}

\item{lon_range}{If not \code{NULL}, a two-element numeric vector indicating latitudinal range of acceptable data. Accepted values are -180 (W) to 180 (E).}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A modified \code{cruz} object. See \code{process_surveys()} documentation for details.
}
\description{
Subsets a \code{cruz} object according to a variety of filters. The same filters
apply to all cohorts within the \code{cruz} object.
}
\details{
This function works by using the filter arguments to determine which rows of \code{DAS} data
meet all criteria; since that \code{DAS} dataframe comes from within the processed \code{cruz} object,
it has a \code{seg_id} column indicating the segment IDs pertaining to eligible rows of data.
Those segment IDs are then used to filter \code{segments}, \code{sightings}, and \code{subgroups}.

Note that the \code{analysis_only} filter is \strong{ignored} for \code{subgroups} sightings/subgroups/events,
since those are often case-by-case-specific.

Note that when arguments are \code{NULL}, they are ignored and no filtering by the associated variable occurs.
Also note that subgroups are not filtered at this time.

Note that there is a similar function, \code{filter_cohort()}, that performs a similar task
for only a single cohort, without requiring an in-tact \code{cruz} object.
Instead it asks for the requisite datasets separately.
}
