% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_combine.R
\name{g0_combine}
\alias{g0_combine}
\title{Combine estimates of weighted g(0) and its CV}
\usage{
g0_combine(g0s, g0cvs)
}
\arguments{
\item{g0s}{A numeric vector of weighted estimates of trackline detection probability, g(0).
Any length is allowed.}

\item{g0cvs}{A numeric vector, the same length as \code{g0s}, of weighted estimates of the CV of g(0).}
}
\value{
A list with two slots: \verb{$g0} holds the combined g(0) estimate, and \verb{$CV} holds the
combined CV.
}
\description{
This function is typically not called by the user, but it certainly can be.
It is called as a subroutine within \code{LTabundR::lta()}.
}
