% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_optimize.R
\name{g0_optimize}
\alias{g0_optimize}
\title{Find g0 values Nelder-Mead using optimization}
\usage{
g0_optimize(g0, g0_cv, try_count = 20, verbose = TRUE)
}
\arguments{
\item{g0}{Initial estimate of g(0).}

\item{g0_cv}{Initial estimate of g(0) coeffiient of variation.}

\item{try_count}{Number of times to attempt optimization before giving up.}

\item{verbose}{Print updates to the Console?}
}
\value{
A list with three named slots: \verb{[[1]] g0_mean}, the exact mean, according to the optimization results;
\verb{[[2]] g0_cv}, the CV according to optimization; and \verb{[[3]] bestFit}, a numeric vector of length 2 containing
parameters to pass to the parameteric g(0) bootstrap routine in \code{LTabundR::lta()}. If the initial g(0) estimate
is 1.0, or the initial g(0) CV estimate is 0.0, then these parameters will be returned as \code{NA}.
}
\description{
This is an internal function used in \code{LTabundR::lta()} for finding the values
(through \code{optim()}) for the parameters needed to conduct parametric bootstrapping
of g0 values for each non-parametric bootstrap of the LTA routine.
}
