% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_plot.R
\name{g0_plot}
\alias{g0_plot}
\title{Plot Rg(0) curves}
\usage{
g0_plot(Rg0, panes = 3, legend_key = 0.4, legend_font = 7)
}
\arguments{
\item{Rg0}{A \code{data.frame} of Rg(0) estimates for a set of species groups,
as produced by the \code{LTabundR} function \code{g0_table()}. See its documentation for details.}

\item{panes}{The number of plot panes to produce. Species will be assigned to each
pane according to their minimum Rg(0) estimate, such that species with similar
detectability will be grouped together.}

\item{legend_key}{Height of each legend key, in cm, to allow for fine-tuning the formatting of the plot.}

\item{legend_font}{Legend font size, to allow for fine-tuning the formatting of the plot.}
}
\value{
A \code{ggplot} object with faceted panes arranged vertically.
}
\description{
This function produces a multi-pane plot of changes in
relative trackline detection probabilities, \emph{Rg(0)},
across Beaufort sea states for a set of species.
}
