% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g0_results}
\alias{g0_results}
\title{Relative trackline detection probabilities, \emph{Rg(0)}}
\format{
An object of class \code{data.frame} with 175 rows and 13 columns.
}
\usage{
g0_results
}
\description{
Relative trackline detection probabilities, \emph{Rg(0)}
}
\details{
\emph{Rg(0)} estimates for 25 species groups, based on survey data from 1986-2020 (see
\code{noaa_10km_1986_2020} dataset).
}
\keyword{datasets}
