% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_table.R
\name{g0_table}
\alias{g0_table}
\title{Compile table of Relative g(0) estimates by Beaufort state}
\usage{
g0_table(
  cruz,
  species,
  eff_types = "S",
  jackknife_fraction = NULL,
  seed = NULL,
  toplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).
Ensure that segment lengths are short; Barlow (2015) used 10km segments.}

\item{species}{A list of sublists, in which each sublist details the settings for
a single species (or related group of species). Each sublist needs to have these slots:
\itemize{
\item \code{spp}: A character vector of species code(s) whose relative trackline detection probability (g(0))
you want to estimate.
\item \code{title}: A unique title, such as a common name, scientific name, or genus,
to easily summarize this species group.
\item \code{cohort}: The cohort whose data pertains to the species of interest,
provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.
If not provided, assumed to be 1.
\item \code{truncation}: The truncation distance, in km, to apply to sightings.
\item \code{constrain_shape} Some \emph{Rg(0)} curves will not decline monotonically due to sample size
issues at low Bft (0-2) or high Bft (5-6) states. To coerce monotonic decline, set this to
\code{TRUE}, and the function will use a shape-constrained GAM (\code{scam()} from package \code{scam}) instead of a
classic \code{mgcv::gam()}.
\item k Smoothing term for the Bft spline in the GAM. Default (and the value used in Barlow 2015) is 4.
\item \code{regions}: A way to specify that \code{cruz} data should be filtered to a certain
geostratum region before conducting the analysis. If not \code{NULL}, this input must match
one of the geostratum names in \code{cruz$strata}.
}}

\item{eff_types}{Effort types to filter segments and sightings to before conducting analysis. The
default is systematic effort only (\code{"S"}). Can be \code{NULL}.}

\item{jackknife_fraction}{The proportion of data to leave out within each jackknife permutation,
which is used for estimating the CV of \emph{Rg(0)} estimates.
The default is 0.1 (i.e., 10\% of the data, yielding 10 jackknife loops), after Barlow (2015).}

\item{seed}{Set a seed (any integer) to ensure that the result is reproducible.
If left \code{NULL}, the results are liable to differ for each run of this function.}

\item{toplot}{Boolean, with default \code{TRUE}, indicating whether segment length histograms and detection function plots (\code{Distance::plot.ds()})
should be displayed (during estimation of effective strip width).}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A \code{data.frame}, in which every row details the Rg0 estimate for a single species group
in a single Beaufort sea state, with the following columns:
\enumerate{
\item \code{title}: The title given to the species group.
\item \code{spp}: Species codes, concatenated by a hyphen if there are multiple in the species group.
\item \code{bft}: Beaufort sea state.
\item \code{Rg0}: Estimate of the relative g(0) for this sea state.
\item \code{ESW}: Estimate of the effective strip half-width (ESW) for this sea state.
\item \code{Rg0_SE}: The standard error of the Rg(0) estimate, non-zero only if \code{jackknife_fraction} was used.
\item \code{Rg0_CV}: The CV of the Rg(0) estimate, non-zero only if \code{jackknife_fraction} was used.
\item \code{ESW_SE}: The standard error of the ESW estimate, non-zero only if \code{jackknife_fraction} was used.
\item \code{sits}: The number of sightings within this sea state used to model Rg(0).
\item \code{sits_p}: The proportion of sightings within this sea state.
\item \code{segs}: The number of segments within this sea state used to model Rg(0).
\item \code{segs_p}: The proportion of segments within this sea state.
}
See the built-in dataset, \code{data(g0_results)}, for an example of the output.
}
\description{
This function estimates relative trackline probabilities, Rg(0),
for a set of species and returns a single table (a \code{data.frame}) with
the results for each species. This function is a wrapper for \code{LTabundR::g0_model()}.
See its documentation for details on the procedure.
}
