% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g0_weighted.R
\name{g0_weighted}
\alias{g0_weighted}
\title{Find a single weighted g(0) estimate (and CV) from a set of Beaufort-specific estimates.}
\usage{
g0_weighted(
  Rg0,
  Rg0_cv,
  cruz,
  cohort = 1,
  iterations = 10000,
  beta_range = c(-1.5, 0),
  beta_step = 0.001,
  beta_sd_range = c(0, 0.3),
  beta_sd_step = 0.0001,
  ymax = 0.2,
  toplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Rg0}{A numeric vector, length 7, of Relative g(0) estimates
for Beaufort sea states 0 - 6. See documentation of \code{LTabundR::g0_bft_model()},
which provides this vector as a column in its output, for details on the
concept of relative trackline detection probabilities.}

\item{Rg0_cv}{A numeric vector, length 7, of the CV of Relative g(0) estimates.
See same documentation as above for details.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort name within your \code{cruz} object that you want to analyze.}

\item{iterations}{The number of draws used to simulate the underlying distribution.}

\item{beta_range}{The range in distribution parameters for simulating the g(0) mean.}

\item{beta_step}{The interval between candidate values within the above range.
A smaller step means higher resolution and accuracy, but takes longer to process.}

\item{beta_sd_range}{The range in distribution parameters for simulating the g(0) SD.}

\item{beta_sd_step}{The interval between candidate values for the above SD range.}

\item{ymax}{The maximum value on the y axis of the diagnostic Rg0 SD plot.}

\item{toplot}{Boolean, with default \code{TRUE}, indicating whether results should be plotted.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
The output is provided in the form of a \code{list} with two slots: \code{g0} and \code{bft}.
The \code{g0} slot holds a \code{data.frame} in which \verb{$wt.mean} is the
weighted mean g(0) value across all Beaufort states for the survey data provided,
and \verb{$wt.cv} is the CV of that estimate.
Other values are reported to facilitate QA/QC, but \verb{$wt.mean} and \verb{$wt.cv} is what
you would pass to an abundance estimation routine, such as \code{LTabundR::lta()}).
The \code{bft} slot holds a dataframe with the proportion of effort in each Beaufort sea state.
}
\description{
Use this function to summarize the Beaufort-specific model of relative trackline probabilities
(from \code{LTabundR::g0_bft_model()}) as a single relative g(0) estimate, weighted
by the prevalence of each Beaufort sea state within the survey data.
This function automatically estimates the CV of this weighted value using an iterative
loop with an MCMC core, adapted from Jeff Moore's code.
}
\details{
This code is an adaptation and expansion of Jeff Moore's code (February 2020).
This function automatically estimates the CV of this weighted value using an iterative
loop with an MCMC core. The MCMC routine is described in Bradford et al. (2021).
The iterative loop tries various candidate values for the distribution parameters and calculates
the cumulative error of the resulting distribution at each Beaufort state. It then
determines the candidate value with the lowest error, weighting each error according to the
proportion of effort occurring in the respective Beaufort state.
}
