% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subgroups.R
\name{get_subgroups}
\alias{get_subgroups}
\title{Find subgroup size estimates within a DAS file}
\usage{
get_subgroups(das, species_filter = "033")
}
\arguments{
\item{das}{A dataframe of \code{DAS} data, formatted by the \code{LTabundR} function \code{load_das()}.}

\item{species_filter}{Species code(s), as a character vector,
if you only want subgroup data for certain species.
The default is for false killer whales, \emph{Pseudorca crassidens}, for whom the subgroup
data collection protocol was developed (see \href{https://www.int-res.com/abstracts/esr/v36/p297-313/]}{Bradford et al. 2018}.}
}
\value{
A dataframe in which each row is a school size estimate for a single subgroup within a single phase of a single sighting
(effectively, the 'raw' data for subgroups within the \code{das} data).
}
\description{
This is an internal function typically not called by a user directly.
It is called in \code{process_subgroups()}, which is an internal function called in \code{process_surveys()}.
}
