% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_effort.R
\name{gg_effort}
\alias{gg_effort}
\title{Add \code{cruz} effort to a base \code{ggplot2} map.}
\usage{
gg_effort(
  p,
  cruz,
  cohort = 1,
  color_by_bft = FALSE,
  color = "black",
  lwd = 0.1,
  alpha = 0.5
)
}
\arguments{
\item{p}{Base \code{ggplot2} to which you are adding this \code{geom}.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{color_by_bft}{Color code by Beaufort sea state?}

\item{color}{Color string.}

\item{lwd}{Effort thickness.}

\item{alpha}{Effort transparency.}
}
\value{
An updated \code{ggplot2} object.
}
\description{
Add \code{cruz} effort to a base \code{ggplot2} map.
}
