% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_geostratum.R
\name{gg_geostratum}
\alias{gg_geostratum}
\title{Add \code{cruz} geostrata to a base \code{ggplot2} map.}
\usage{
gg_geostratum(
  p,
  cruz,
  strata,
  color = "orchid4",
  lwd = 0.5,
  lty = 1,
  crs = 4326
)
}
\arguments{
\item{p}{Base \code{ggplot2} to which you are adding this \code{geom}.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{strata}{desc}

\item{color}{Color string.}

\item{lwd}{Boundary line width.}

\item{lty}{Boundary line type.}

\item{crs}{Optional CRS.}
}
\value{
An updated \code{ggplot2} object.
}
\description{
Add \code{cruz} geostrata to a base \code{ggplot2} map.
}
