% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_sightings.R
\name{gg_sightings}
\alias{gg_sightings}
\title{Add \code{cruz} sightings to a base \code{ggplot2} map.}
\usage{
gg_sightings(
  p,
  cruz,
  cohort = 1,
  spp = NULL,
  spp_translate = NULL,
  color_by_spp = FALSE,
  color = "black",
  pch = 16,
  cex = 0.5,
  alpha = 0.5
)
}
\arguments{
\item{p}{Base \code{ggplot2} to which you are adding this \code{geom}.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{spp}{Character vector of species codes.}

\item{spp_translate}{Dataframe of scecies code information for translating between codes and common names. See \code{data(species_codes)} for an example.}

\item{color_by_spp}{Color code by species code?}

\item{color}{Color string.}

\item{pch}{Sighting point type.}

\item{cex}{Sighting point size.}

\item{alpha}{Sighting transparency.}
}
\value{
An updated \code{ggplot2} object.
}
\description{
Add \code{cruz} sightings to a base \code{ggplot2} map.
}
