% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_idl.R
\name{ggplot_idl}
\alias{ggplot_idl}
\title{Modify a \code{ggplot} object to handle the International Date Line (IDL)}
\usage{
ggplot_idl(
  p_base,
  lon_range = c(100, -100),
  lat_range = NULL,
  idl_padding = 0.02,
  axes = TRUE,
  bypass = FALSE
)
}
\arguments{
\item{p_base}{A \code{ggplot} object that you want to visualize in a way that handles the International Date Line.}

\item{lon_range}{Desired longitude range, as a two-element vector.
The first element should be a positive number between 0 and 180 (indicating decimal degrees East),
the second should be a negative number between -180 and 0 (indicating decimal degrees West).}

\item{lat_range}{Desired latitude range, as a two-element vector of decimal degrees (positive = N, negative = S).}

\item{idl_padding}{The space between the Eastern hemisphere plot and the Western hemisphere plot.}

\item{axes}{If \code{FALSE}, axis labels and ticks will be removed.}

\item{bypass}{An option to return an unmodified \code{ggplot}, to ease automated toggling when using in various functions within \code{LTabundR}.}
}
\value{
A \code{ggplot2} object that is actually two plots side by side, with a small space in between indicating the IDL.
}
\description{
Modify a \code{ggplot} object to handle the International Date Line (IDL)
}
