% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_size.R
\name{group_size}
\alias{group_size}
\title{Group size estimation wrapper}
\usage{
group_size(
  grp,
  gs_coefficients = NULL,
  calibrate_floor = 0,
  geometric_mean = FALSE,
  use_low_if_na = TRUE,
  debug_mode = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{grp}{A dataframe, in which each row is a school size estimate from a single observer,
and each column is a column from the \code{DAS} dataframe that is relevant to school size estimation
(columns Event, year, Bft, and Prob:GsSchoolLow).}

\item{gs_coefficients}{If not \code{NULL}, a dataframe of school size calibration
coefficients (see description in \code{load_survey_settings()}).}

\item{calibrate_floor}{This argument accepts a number
indicating the minimum raw school size estimate
for which school size calibration will be attempted.
When this function is used withing \code{process_sightings()},
this setting from the \code{cruz} object will be provided.}

\item{geometric_mean}{This argument accepts a Boolean;
if \code{TRUE}, geometric means will be calculated instead of arithmetic means.
If school size calibration is carried out,
the geometric mean will be weighted by calibration variance, such that
estimates from observers with low variance will receive more weight.
When this function is used withing \code{process_sightings()},
this setting from the \code{cruz} object will be provided.}

\item{use_low_if_na}{If this setting is \code{TRUE},
when no observer makes a best estimate of group size,
mean group size will be calculated from "low" estimates.
This will be done only if no observer has a "best" estimate.
When this function is used withing \code{process_sightings()},
this setting from the \code{cruz} object will be provided.}

\item{debug_mode}{Boolean, with default \code{FALSE},
indicating whether details should be printed to the Console that facilitate debugging.}

\item{verbose}{Boolean, with default \code{TRUE},
indicating whether or not updates should be printed to the Console.}
}
\value{
A dataframe in which each row is a species within the sighting,
with final estimates of best / high / low and metadata regarding calibration.
Passed back to \code{process_sightings()}.
}
\description{
This is an internal function typically not called by a user directly.
This function is used as a subroutine in \code{process_sightings()}.
It produces school size estimates for each species within a sighting,
based upon all estimates of school size and species composition provided by observers.
}
