% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_calibrate_abund.R
\name{grp_calibrate_abund}
\alias{grp_calibrate_abund}
\title{Group size calibration}
\usage{
grp_calibrate_abund(
  obs,
  bft,
  yr,
  gbest,
  glow,
  ghigh,
  gs_coefficients = NULL,
  calibrate_floor = 0
)
}
\arguments{
\item{obs}{Character vector with single observer ID. If this observer ID is contained
within \code{gs_coefficients}, its specific calibration coefficients will be used.
Otherwise the generic calibration coefficients will be used.}

\item{bft}{Numeric of Beaufort sea state during sighting.}

\item{yr}{Numeric of year of sighting.}

\item{gbest}{Numeric of best school size estimate.}

\item{glow}{Numeric of low school size estimate.}

\item{ghigh}{Numeric of high school size estimate.}

\item{gs_coefficients}{Dataframe of school size coefficients. If not provided, calibration will not be attempted.}

\item{calibrate_floor}{This argument accepts a number indicating the minimum raw school size estimate
for which school size calibration will be attempted.}
}
\value{
A one-row dataframe with the best school size estimate, the variance of the estimate,
a boolean indicating whether calibration was possible, and the slope and intercept used for calibration.
Passed back to \code{group_size()}.
}
\description{
This is an internal function typically not called by a user directly.
It is called by the subroutine \code{group_size()}, which is itself a subroutine
of \code{process_sightings()}.
\cr \cr
This function performs school size calibration for a single observer's estimates.
}
