% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grp_coeff_abund}
\alias{grp_coeff_abund}
\title{Group size coefficients for the ABUND calibration approach}
\format{
An object of class \code{data.frame} with 162 rows and 21 columns.
}
\usage{
grp_coeff_abund
}
\description{
Group size coefficients for the ABUND calibration approach
}
\details{
A table of group size calibration coefficient values for a variety of species
for a variety of observer codes. This table was provided by Jay Barlow (NOAA NMFS biologist), was created
prior to 2016, and has been used for published analyses through 2021 with ABUND versions 7 & 8.
}
\keyword{datasets}
