% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grp_coeff_gerrodette}
\alias{grp_coeff_gerrodette}
\title{Group size coefficients for the Gerrodette calibration approach}
\format{
An object of class \code{data.frame} with 19 rows and 4 columns.
}
\usage{
grp_coeff_gerrodette
}
\description{
Group size coefficients for the Gerrodette calibration approach
}
\details{
A table of group size calibration coefficient values for a variety of species
for use in the "Gerrodette" calibration approach. These values were drawn
from Gerrodette et al. 2019 and used in Barlow et al. (2026).
}
\keyword{datasets}
