% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grp_ops_abund}
\alias{grp_ops_abund}
\title{Calibration options for ABUND approach}
\format{
An object of class \code{list} of length 6.
}
\usage{
grp_ops_abund
}
\description{
Calibration options for ABUND approach
}
\details{
Example of calibration options to use for the ABUND group-size-calibration approach.
}
\keyword{datasets}
