% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grp_ops_gerrodette}
\alias{grp_ops_gerrodette}
\title{Calibration options for Gerrodette approach}
\format{
An object of class \code{list} of length 6.
}
\usage{
grp_ops_gerrodette
}
\description{
Calibration options for Gerrodette approach
}
\details{
Example of calibration options to use for the "Gerrodette" group-size-calibration approach.
}
\keyword{datasets}
