% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_recalibrate.R
\name{grp_recalibrate}
\alias{grp_recalibrate}
\title{Recalibrate processed sightings}
\usage{
grp_recalibrate(
  cruz,
  cohort = 1,
  spp = NULL,
  include_mixed = TRUE,
  mixed_only = FALSE,
  spp_max_only = FALSE,
  ops_max = grp_ops_gerrodette,
  ops_not = grp_ops_abund,
  verbose = TRUE,
  toplot = TRUE
)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose sightings you would like to recalibrate,
provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{spp}{Character vector of species codes to filter by.
Leave \code{NULL} if you want to recalibrate all sightings in the cohort.}

\item{include_mixed}{If \code{TRUE}, mixed-species groups containing the \code{spp}
of interest will be recalibrated, including estimates for non-target species
present in those mixed groups.}

\item{mixed_only}{If \code{TRUE}, \emph{only} mixed-species groups will be recalibrated.
Ignored if \code{include_mixed} is \code{FALSE}.}

\item{spp_max_only}{If \code{TRUE}, the only mixed-species groups recalibrated
will be those in which the \code{spp} of interest are the most abundant species
in the group (i.e., \code{spp_max} is found in \code{spp}).  Ignored if \code{include_mixed} is \code{FALSE}.}

\item{ops_max}{The calibration options to apply to the predominant species
in the group. Default is the "Gerrodette" method, with default values provided
in the built-in dataset \code{data(grp_ops_gerrodette)}.}

\item{ops_not}{The calibration options to apply to the non-predominant species
in any mixed groups contained in the set of sightings to be recalibrated.
Default is the "ABUND" method, with default values provided
in the built-in dataset \code{data(grp_ops_abund)}.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}

\item{toplot}{Boolean, with default \code{TRUE}, indicating whether a diagnostics plot should be displayed at the end of the recalibration.}
}
\value{
A \code{list} with the following slots:
\enumerate{
\item \code{key}: A \code{data.frame} key of which sightings were re-calibrated,
with three columns: \code{i} = row index from \code{cruz$cohorts[[cohort]]$sightings};
\code{SightNoDaily} = unique Sighting identifier from same;
\code{species} = the species code.
\item \code{recalibrated_sits}: Complete \code{sightings} rows but only for the sightings
that were re-calibrated.
\item \code{updated_cruz}: Updated \code{cruz} object in which \code{cruz$cohorts[[cohort]]$sightings}
includes the recalibrated sightings as well as all untouched sightings.
}
}
\description{
Re-run group size calibration for a selection of sightings within a pre-existing \code{cruz} object.
}
