% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_size.R
\name{grp_size}
\alias{grp_size}
\title{Group size estimation}
\usage{
grp_size(
  grp,
  calibrate = grp_ops_abund,
  geometric_mean = FALSE,
  use_low_if_na = TRUE,
  debug_mode = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{grp}{A dataframe, in which each row is a school size estimate from a single observer,
and each column is a column from the \code{DAS} dataframe that is relevant to school size estimation
(columns Event, year, Bft, and Prob:GsSchoolLow).}

\item{calibrate}{If \code{NULL}, group sizes will not be calibrated.
Otherwise supply a \code{list} to specify which calibration approach to use.
Currently two approaches are supported: that used by \verb{ABUND 7/8} (the default shown),
which takes an observer-specific, species-agnostic approach,
and that developed in Gerrodette et al. (2019) and applied in Barlow et al. (2026),
which includes a species-specific, observer-agnostic approach. Details in \code{?load_cohort_settings()}.}

\item{geometric_mean}{This argument accepts a Boolean;
if \code{TRUE}, geometric means will be calculated instead of arithmetic means.
If school size calibration is carried out,
the geometric mean will be weighted by calibration variance, such that
estimates from observers with low variance will receive more weight.
When this function is used withing \code{process_sightings()},
this setting from the \code{cruz} object will be provided.
Note that, although only the best estimates may be calibrated if specified above (never the highs and lows),
the \emph{same} kind of averaging function \emph{is} applied to the highs and lows as is applied to the bests.
That is, when \code{geometric_mean} is \code{TRUE}, the geometric mean of the highs and the lows is returned.
If the best estimates are calibrated, the geometric \emph{weighted} mean will be applied to the highs and lows,
using the variance of the calibrated best estimates as weights. If the best estimates are \emph{not} calibrated,
the \emph{unweighted} geometric mean is used to estimate the highs, lows, and bests.}

\item{use_low_if_na}{If this setting is \code{TRUE},
when no observer makes a best estimate of group size,
mean group size will be calculated from "low" estimates.
This will be done only if no observer has a "best" estimate.
When this function is used withing \code{process_sightings()},
this setting from the \code{cruz} object will be provided.}

\item{debug_mode}{Boolean, with default \code{FALSE},
indicating whether details should be printed to the Console that facilitate debugging.}

\item{verbose}{Boolean, with default \code{TRUE},
indicating whether or not updates should be printed to the Console.}
}
\value{
A dataframe in which each row is a species within the sighting,
with final estimates of best / high / low and metadata regarding calibration.
When this function is used internally by \code{LTabundR}, the results are passed back to \code{process_sightings()}.
}
\description{
This is an internal function typically not called by a user directly.
This function is used as a subroutine in \code{process_sightings()}.
It produces school size estimates for each species within a sighting,
based upon all estimates of school size and species composition provided by observers.
}
