% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idl_coerce.R
\name{idl_coerce}
\alias{idl_coerce}
\title{Coerce eastern longitudes to western (i.e., less than - 180)}
\usage{
idl_coerce(df)
}
\arguments{
\item{df}{Dataframe of coordinates with columns \code{Lon} and \code{Lat}.}
}
\value{
Corrected dataframe (all coordinates between 0 and less than -180).
}
\description{
This can be useful for plotting datasets that cross the international date line (IDL).
(Designed as a behind-the-scences function that most users wont' use, but making it available just in case.)
}
