% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Land}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 10 rows and 4 columns.
}
\usage{
land
}
\description{
Land
}
\details{
World dataset of land polygons, downloaded from Natural Earth.
}
\keyword{datasets}
