% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cohort_settings.R
\name{load_cohort_settings}
\alias{load_cohort_settings}
\title{Load cohort-specific settings}
\usage{
load_cohort_settings(
  id = "default",
  species = NULL,
  strata = NULL,
  probable_species = FALSE,
  sighting_method = 0,
  cue_range = 0:7,
  group_size_range = c(0, 10000),
  group_size_calibrate = grp_ops_abund,
  use_low_if_na = FALSE,
  io_sightings = 0,
  geometric_mean_group = TRUE,
  truncation_km = 20,
  beaufort_range = 0:6,
  abeam_sightings = FALSE,
  strata_overlap_handling = "smallest",
  distance_types = c("S", "F", "N"),
  distance_modes = c("P", "C"),
  distance_on_off = TRUE
)
}
\arguments{
\item{id}{An informal identifier for this cohort,
to help you keep track of which cohort is which.
For example, settings for a cohort of large whales species could be named \code{"big_whales"};
settings for small delphinids and phocoenids could be named \code{"small_odontocetes"};
settings for beaked whales could be named \code{"beakers"}.}

\item{species}{A character vector of species codes to include in this cohort.
If \code{NULL} (the default), all species will be included.}

\item{strata}{A character vector of geostratum names; these must match the names
listed in the \code{strata} slot of your survey settings (see documentation for \code{load_settings()}).
If \code{NULL} (the default), \emph{all} geostrata in your survey settings will each be used.
This argument is an opportunity to subset the geostrata used for a cohort;
for example, certain dolphin species in Hawaiian waters have unique geostrata that
should apply to them but not to other species.}

\item{probable_species}{If \code{TRUE} (default is \code{FALSE}),
the “probable” species identifications will be used in place of the “unidentified” categories.}

\item{sighting_method}{A coded integer which determines which sightings
will be included based on how they were first seen.
Allowable codes are
\code{0}=any method,
\code{1}=with 25X only,
\code{2}=neither with 25x binoculars nor from the helicopter (i.e., naked eyes and 7x binoculars only).
These codes match those used in \code{ABUND7/9}.}

\item{cue_range}{Numeric vector of acceptable "observation cues"
for sightings used in estimates of abundance.
(\code{0}=this detail is missing in the data,
\code{1}=associated birds,
\code{2}=splashes,
\code{3}=body of the marine mammal,
\code{4}=associated vessel,
\code{5}=?,
\code{6}=blow / spout,
\code{7}=associated helicopter).
These codes match those used in \code{ABUND7/8/9}.}

\item{group_size_range}{Minimum and maximum group sizes to be included in estimates of abundance.
This is the overall group size, not the number of the given species that are present in a mixed group.}

\item{group_size_calibrate}{If \code{NULL}, group sizes will not be calibrated.
Otherwise supply a \code{list} to specify which calibration approach to use.
Currently two approaches are supported: that used by \verb{ABUND 7/8} (the default shown),
which takes an observer-specific, species-agnostic approach,
and that developed in Gerrodette et al. (2019) and applied in Barlow et al. (2026),
which includes a species-specific, observer-agnostic approach. See Details below.}

\item{use_low_if_na}{If this setting is \code{TRUE},
and if an observer does not make a best estimate of group size,
mean group size will be calculated from "low" estimates.
This will be done only if no observer has a "best" estimate.}

\item{io_sightings}{A coded integer which specifies how sightings by the independent observer will be handled.
Allowable codes, which are inherited from those used in \code{ABUND7/9}, are
\code{"-1"}=include independent observer sightings with all other sightings,
\code{"0"}=ignore sightings by independent observer,
\code{"1"}=use only sightings made by regular observer team WHEN an independent observer was present,
\code{"2"}=include only sightings made by the independent observer.
IO sightings are typically used only for making g(0) estimates,
otherwise IO sightings are usually ignored (code = \code{"0"}).}

\item{geometric_mean_group}{This argument accepts a Boolean;
if \code{TRUE}, geometric mean school sizes will be calculated instead of arithmetic means.
Also, if school size calibration is carried out, the geometric mean will be weighted by calibration variance,
such that estimates from observers with low variance will receive more weight.
Barlow, Gerrodette, and Perryman (1998) found that using the geometric mean
yielded slightly better performance than a simple arithmetic mean group size.}

\item{truncation_km}{Specifies the maximum perpendicular distance
for groups that are to be included for abundance estimation.
Also determines the bins used for grouped perpendicular distances.}

\item{beaufort_range}{A numeric vector indicating the Beaufort sea states (0 - 7) to be accepted within usable segments.}

\item{abeam_sightings}{If \code{TRUE}, sightings that occur aft of beam are included
in estimating the detection function and densities. Default is \code{FALSE}:
all abeam sightings will not be used in density estimation or detection function estimation.}

\item{strata_overlap_handling}{This setting informs how effort is split into
segments when surveys cross stratum boundaries, and also which stratum name
is assigned to each row of data. See Details below.}

\item{distance_types}{A character vector of the full-range of effort types that meet the "analysis
inclusion criteria", i.e., will be included in detection function estimation,
and therefore considered in effort segmentizing.
Accepted values are \code{"S"} (systematic/standard effort), \code{"F"} (fine-scale effort),
and \code{"N"} (non-systematic/non-standard effort, in which systematic protocols
are being used but effort is not occurring along design-based transect routes).
Note that it is possible to further subset the effort types specifically for
density estimation in \code{LTabundR}'s function for line transect analysis, \code{lta()}.}

\item{distance_modes}{The effort modes that meet "analysis inclusion criteria", i.e.,
will be included in detection function estimation,
and therefore considered in effort segmentizing.
Accepted values are \code{"P"} (passing) and \code{"C"} (closing)}

\item{distance_on_off}{The value(s) of \code{OnEffort}
(On Effort is \code{TRUE}, Off Effort is \code{FALSE}) that will be included in
detection function estimation,
and therefore considered in effort segmentizing.}
}
\value{
A list with named slots, equivalent to your input arguments.
Save this output to an object, e.g., using the same name you provided in the \code{id} argument,
and pass it to \code{load_settings()}.
}
\description{
This function builds a list of cohort-specific settings -- applying only to the analyses
for a certain group of species -- which you will pass to \code{load_settings()}.
}
\details{
More information on the \code{list} required for the \strong{\code{group_size_calibrate}} input:
The \code{list} requires the following fields (examples are given in the built-in datasets
\code{grp_ops_abund} and \code{grp_ops_gerrodette}):
\itemize{
\item \code{method}: accepted values = 'ABUND' and 'Gerrodette';

\item \code{coefficients}: a \code{data.frame}. For the "ABUND" method, this is a table of observer-specific calibration coefficients.
To use the same coefficients that have been in use at SWFSC and PIFSC up to 2021,
see the built-in dataset \code{data(grp_coeff_abund)}.
For the "Gerrodette" method, this is a table of species-specific coefficients.
To use the same coefficients from the method's originating papers, see the
built-in dataset \code{data(grp_coeff_gerrodette)}.
For both methods, supplied \code{data.frame}'s must match the column naming structure of the respective built-in dataset.

\item \code{floor}: minimum group size to which calibration will be applied.
In the "ABUND" approach, this is referenced only for observers who do not
have an entry in the \code{coefficients} table, which has a calibration floor for each observer.
The default is 0, meaning that calibration will be attempted for \emph{all}
school size estimates, regarding of the raw estimate.

\item \code{intercept} (ignored if not using the "Gerrodette" method) the value to
use for the "intercept" coefficient if a species in this cohort is not present in the \code{coefficients} table.

\item \code{beta} (ignored if not using the "Gerrodette" method): the value to
use for the "beta" coefficient if a species in this cohort is not present in the \code{coefficients} table.

\item \code{beta_mixed} (ignored if not using the "Gerrodette" method): used for mixed-species groups.
If \code{NULL}, the "beta" coefficient for the most abundant species in the group will be used.
If a numeric value is supplied here, that value will be used instead.
}
Note that the high and low estimates are \emph{never} calibrated; only the best estimates are.

More information on the \code{list} required for the \strong{\code{strata_overlap_handling}} input:
Note that the main impact of this setting is
on how effort is broken into segments; the assigned stratum name is for display
only and will not constrain options for including/excluding strata in analyses
farther along in the \code{LTabundR} workflow.

\itemize{
\item The default option is \code{"smallest"},
which means that effort will always be assigned to the smallest stratum when
multiple strata overlap spatially. This is a safe option for surveys with "nested"
strata (such as the Central North Pacific strata used by NOAA Fisheries.

\item Another option is \code{"each"}in which each time a stratum boundary is crossed the
current segment will end and a new segment will begin.
Also, stratum assignments for each row of effort will be shown as a concatenation
of all the stratum layers overlapping at its position (e.g., "OtherCNP&HI_EEZ").
Note that the \code{"each"} option segmentizes effort in the exact same was as \code{"smallest"}
when strata are fully nested; its main advantage is in dealing with partially
overlapping strata.

\item The third option is \code{"largest"}, in which the largest of
overlapping strata is used to assign a stratum name to each row.
(We are not sure what use case this would serve,
but we offer it as an option for niche analyses.)
}
}
