% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_checks.R
\name{lta_checks}
\alias{lta_checks}
\title{Check for missing data before running \code{lta()}.}
\usage{
lta_checks(cruz, df_settings, fit_filters, estimates)
}
\arguments{
\item{cruz}{The same \code{cruz} object you will pass to \code{lta()}. See input details in that function's documentation.}

\item{df_settings}{The same \code{df_settings} object you will pass to \code{lta()}. See input details in that function's documentation.}

\item{fit_filters}{The same \code{fit_filters} object you will pass to \code{lta()}. See input details in that function's documentation.}

\item{estimates}{The same \code{estimates} object you will pass to \code{lta()}. See input details in that function's documentation.}
}
\value{
The function prints messages to the Console as it performs its checks,
and it also returns a \code{list} with all details needed to find & edit rows in the sightings data that have missing data.
}
\description{
This function is designed for a QA/QC step before running the \code{lta()} function
(line transect analysis). It checks for missing data in covariate columns,
group size columns (if \code{lnsstot} is a desired covariate), and the perpendicular distance column (\code{PerpDistKm}).
The function runs these checks twice: first for all sightings to be used in fitting the detection function
(where missing data are not preferred but would not be fatal; sightings with incomplete data would simply be removed,
lowering your sample size for the model fit), and second for detections that contribute to each point estimate
(here missing data is disastrous; unless the missing data are filled in, the detection would be removed and the density estimate would be affected.)
This function can be used to pinpoint and address holes in the data as the user sees fit.
}
