% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_ci.R
\name{lta_ci}
\alias{lta_ci}
\title{Compute confidence intervals of a bootstrapped LTA estimate}
\usage{
lta_ci(estimate, bootstraps, ci = 0.95)
}
\arguments{
\item{estimate}{The actual estimate}

\item{bootstraps}{The bootstrapped values}

\item{ci}{The percent confidence interval sought (as a decimal).}
}
\value{
A \code{list} with three versions of the 95\% confidence interval,
in each case returned as a two-element vector: \verb{$quantile} is the simple quantile-based
CI of the bootstraps; \verb{$bca} is the bias-correction with acceleration; and \verb{$bca_lognormal}
is the log-normal version of the \code{bca} CI. Adapted from Karin Forney's Excel magic.
}
\description{
Compute confidence intervals of a bootstrapped LTA estimate
}
