% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_combine.R
\name{lta_combine}
\alias{lta_combine}
\title{Combine LTA estimates from separate regions}
\usage{
lta_combine(d1, n1, cv1, lci1, uci1, area1, d2, n2, cv2, lci2, uci2, area2)
}
\arguments{
\item{d1}{Density estimate (animals per square km) from region 1.}

\item{n1}{Abundance estimate from region 1.}

\item{cv1}{CV estimate from region 1.}

\item{lci1}{Lower 95\% confidence interval of abundance from region 1.}

\item{uci1}{Upper 95\% confidence interval of abundance from region 1.}

\item{area1}{Area, in square km, of region 1.}

\item{d2}{Density estimate for region 2.}

\item{n2}{Abundance estimate for region 2.}

\item{cv2}{CV estimate for region 2.}

\item{lci2}{LCI for region 2.}

\item{uci2}{UCI for region 2.}

\item{area2}{Area for region 2.}
}
\value{
A one-row \code{data.frame} with these columns:
\enumerate{
\item \code{D}: the combined density (animals per square km), weighted by the areas of constituent regions.
\item \code{N}: the combined abundance.
\item \code{CV}: the combined CV of density and abundance.
\item \code{LCI}: the lower 95\% confidence interval of abundance.
\item \code{UCI}: the upper 95\% confidence interval of abundance.
\item \code{area}: combined area.
}
}
\description{
Combine the results of \code{LTabundR::lta()} from separate regions,
weighting density by respective area and
handling the combination of CV and 95\% confidence interval estimates.
}
