% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_delta_method.R
\name{lta_delta_method}
\alias{lta_delta_method}
\title{Estimate CV and 95\% confidence intervals with the delta method}
\usage{
lta_delta_method(ltas, verbose = TRUE)
}
\arguments{
\item{ltas}{A \code{list} of outputs from the function \code{lta()}.}

\item{verbose}{Boolean; print report of new estimates to the console?}
}
\value{
An updated version of the \code{ltas} input with new values for \code{CV}, \code{L95}, and \code{U95} in the \verb{$bootstrap$summary} slot.
}
\description{
This function loops through a list of results from \code{LTabundR::lta()} and updates
their CV and 95\% CI estimates, which were produced using a bootstrap method,
using new estimates based on the delta method.
}
