% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_destratify.R
\name{lta_destratify}
\alias{lta_destratify}
\title{Pool stratified LTA results}
\usage{
lta_destratify(
  lta_list,
  years,
  combine_method = "arithmetic",
  new_region = NULL,
  verbose = TRUE
)
}
\arguments{
\item{lta_list}{A \code{list} of line-transect-analysis results produced from
\code{LTabundR::lta()} and compiled into a \code{list} with \code{LTabundR::lta_enlist()}.}

\item{years}{A numeric vector of year(s) in which you want to pool all region-stratified
results into a single annual estimate.}

\item{combine_method}{A character string indicating the means by which estimates
of the CV and 95\% confidence interval of abundance will be destratified.
The two recognized options are \code{"arithmetic"} (the default) and \code{"bootstrap"}.
See details below.}

\item{new_region}{A name for the pooled geostratum represented by the pooled estimate.
If this is not provided, a name will be generated from the geostrata that are pooled.}

\item{verbose}{A Boolean, with default \code{TRUE}, indicating whether or not status
updates should be printed to the Console.}
}
\value{
A modified \code{lta_list}, in which stratified estimates have been pooled into
single annual estimates for relevant years.
}
\description{
Loop through a list of stratified line-transect-analysis results
and "destratify" -- i.e., pool -- them into a single estimate.
}
\details{
Stratified estimates of parameter means are pooled using a weighted mean
in which the area of each geostratum is used as the weight. Stratified estimates of
abundance variance terms (SD, CV, and 95\% confidence interval) are pooled either
arithmetically -- using standard equations to pool estimates one at a time --
or iteratively -- using the bootstrap results contained within the \code{lta_list}.
The iterative technique resamples the abundance bootstrap estimates from each
geostratum 10,000 times then calculates SD, CV, and 95\% confidence interval from the
resulting distribution. Some variance terms, e.g., the SD of school size and the CV of g(0),
are not destratified in the current version and are instead replaced with \code{NA}.
}
