% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_diagnostics.R
\name{lta_diagnostics}
\alias{lta_diagnostics}
\title{Run diagnostics on a LTA result}
\usage{
lta_diagnostics(lta_result)
}
\arguments{
\item{lta_result}{The result of \code{lta()} for a single species pool.}
}
\value{
A series of printed tables and plots, which the user can step through using the \code{Enter} key.
}
\description{
Run diagnostics on a LTA result
}
