% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_enlist.R
\name{lta_enlist}
\alias{lta_enlist}
\title{Compile a directory of LTA results into a list}
\usage{
lta_enlist(lta_path)
}
\arguments{
\item{lta_path}{The path to a directory containing \code{.RData} files, each containing a object
resulting from \code{LTabundR::lta()}. Typically, each \code{RData} file contains the product of \code{lta()} for a
single species or species pool. See the \href{https://emk-noaa.github.io/LTAvignette/whiceas.html#density-abundance}{vignette case studies} for examples.}
}
\value{
A \code{list} of LTA results.
}
\description{
This function reads in the line-transect-analysis results (from \code{LTabundR::lta()}) within a directory
and concatenates them into a \code{list}.
}
