% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_estimates.R
\name{lta_estimates}
\alias{lta_estimates}
\title{Build a function that creates \code{estimates} sub-lists for \code{lta()}}
\usage{
lta_estimates(scenarios)
}
\arguments{
\item{scenarios}{A list of sub-lists. Each sublist specifies one scenario for a density/abundance estimate.
The accepted names within each sublist are:
\itemize{
\item \code{years}
\item \code{cruises}
\item \code{regions}
\item \code{regions_remove}
\item \code{region_title}
\item \code{forced_effort}
\item \code{area}
\item \code{remove_land}
}
These arguments are explained under the \code{estimates} input for the \code{lta()} function.}
}
\value{
This function returns a custom function named \code{estimator()}, which accepts the following inputs:
\itemize{
\item \code{spp} (required)
\item \code{title} (required)
\item \code{g0} (the remainder are optional)
\item \code{g0_cv}
\item \code{g0_threshold}
\item \code{alt_g0_spp}
\item \code{combine_g0}
}
These arguments are explained under the \code{estimates} input for the \code{lta()} function.
The \code{estimator()} function will return a list of sub-lists, in which the
species-specific inputs it accepts are added to each sub-list in the \code{scenarios} argument you provided above.
The result of the \code{estimator()} function is what you can pass as the \code{estimates} argument in \code{lta()}.
See the \href{https://emk-noaa.github.io/LTAvignette/whiceas.html#density-abundance}{vignette case studies} for examples.
}
\description{
This function can be used to efficiently build up sub-lists for the \code{estimates} input
that you must pass to the \code{lta()} function for line-transect analysis.
Most studies of multiple species endeavor to estimate density/abundance for each species
in the same set of year-region scenarios.
Doing so typically requires multiple \code{estimates} sub-lists for each species of interest.
Writing those sub-lists manually can be redundant, tedious, and prone to inconsistencies or error.
To ensure the exact same scenarios are applied to each species of interest, you
can use this function to return a custom function that you can then use to creates a
standard set of sub-lists for each of you species.
}
