% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_plot.R
\name{lta_plot}
\alias{lta_plot}
\title{Plot LTA results}
\usage{
lta_plot(lta_result, species = NULL, years = NULL, nrows = NULL, ncols = NULL)
}
\arguments{
\item{lta_result}{The result of \code{lta()} or \code{lta_enlist()}. See their documentation for details.}

\item{species}{A character vector of estimate titles. If \code{NULL}, all titles within \code{lta_result} will be used.}

\item{years}{A numeric vector of years whose estimates will be included in all plots,
even if an estimate is not available for a given \code{species} (to make the x-ranges of all plots equal)}

\item{nrows}{Number of rows in multi-panel plot. Can be \code{NULL} (which is the default).}

\item{ncols}{Number of columns in multi-panel plot. Can be \code{NULL} (which is the default).}
}
\value{
A \code{ggplot2} plot, or many \code{ggplot2} plots arranged by \code{gridExtra::grid.arrange()}.
}
\description{
This function returns abundance estimate plots (i.e., with year on the x axis,
abundance on the y axis, the best estimate represented by a point, and the confidence
interval represented by a vertical line). This function is capable of producing multi-panel
plots for multiple species, and it is designed to produce a dynamic layout according to the
number of species pools provided.
}
