% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_pool.R
\name{lta_pool}
\alias{lta_pool}
\title{Pool bootstraps from multiple \code{lta()} outputs}
\usage{
lta_pool(ltas, bootstraps = NULL)
}
\arguments{
\item{ltas}{A \code{list} of outputs from the function \code{lta()}.}

\item{bootstraps}{Desired number of bootstraps to keep. Example use case: You have two \code{lta()} outputs,
one with 600 iterations and another with 500. You want to pool the two outputs but only keep 1,000 iterations total.
Specify \code{bootstraps = 1000} and this function will randomly select 1,000 of the 1,100 bootstraps available.}
}
\value{
An \code{lta()} output object.
The \verb{$estimate} slot will be the exact same as the first slot in the \code{ltas} input;
The \verb{$details} slot will have the pooled bootstrap samples.
The \verb{$df} slot will have the pooled detection function curves.
The \verb{$summary} slot will have updated summary statistics based on the pooled set of bootstraps.
}
\description{
This function allows you to pool, or concatenate, the results of two separate \code{lta()} runs
in order to achieve the desired number of bootstrap iterations. An example use case:
you are running a 1,000-iteration \code{lta()} call overnight, but the power goes out at iteration 900.
In the morning you can run a 100-iteration version of the same \code{lta()} call, then use
this function to pool the two results together.
}
