% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lta_result}
\alias{lta_result}
\title{Example of LTA results for Striped Dolphin, 2010 and 2017}
\format{
An object of class \code{list} of length 5.
}
\usage{
lta_result
}
\description{
Example of LTA results for Striped Dolphin, 2010 and 2017
}
\details{
Details to come.
}
\keyword{datasets}
