% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_base.R
\name{map_base}
\alias{map_base}
\title{Base map for plotting survey data}
\usage{
map_base(
  region = "cnp",
  eez_color = "grey90",
  eez_stroke = 2,
  land_color = "grey60",
  land_stroke = 0.5,
  xlims = NULL,
  ylims = NULL,
  scale_max = 1000
)
}
\arguments{
\item{region}{Name of the survey region you wish to map.
Options currently supported:
\code{"cnp"} (the default; Central North Pacific), \code{"ccs"} (California Current System),
and \code{"etp"} (Eastern Tropical Pacific).
This argument will define the default bounding box of the map and the EEZ's that are plotted.}

\item{eez_color}{Color of the EEZ boundary lines.}

\item{eez_stroke}{Width of the EEZ boundary lines.}

\item{land_color}{Color of land in the map.}

\item{land_stroke}{Width of the borders of land.}

\item{xlims}{Custom longitude limits, provided as a two-element numeric vector of decimal degrees with West coordinates negative; useful if you wish to use a "zoom" different from the default set by \code{region}.}

\item{ylims}{Custom latitude limits (South coordinates are negative).}

\item{scale_max}{Distance (in km) to be represented by the scale.}
}
\value{
A \code{tmap} object (the map is built using the package \code{tmap}), which will print as a plot if not saved as an object.
If saved as an object, other features can be added to the map (see \code{map_effort()}, \code{map_strata()}, and \code{map_sightings()},
or other \code{tmap} functions).
}
\description{
Use this function io initiate a map of your survey.
}
