% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cruz.R
\name{map_cruz}
\alias{map_cruz}
\title{Interactive map for \code{Wincruz} survey}
\usage{
map_cruz(
  cruz,
  cohort = 1,
  eez_show = TRUE,
  eez_color = "black",
  eez_weight = 0.6,
  eez_opacity = 0.5,
  strata_show = TRUE,
  strata_color = "forestgreen",
  strata_weight = 1,
  strata_opacity = 0.5,
  effort_show = FALSE,
  effort_resolution = 15,
  effort_color = "darkblue",
  effort_weight = 0.6,
  effort_opacity = 0.85,
  sightings_show = TRUE,
  sightings_color = "color code",
  sightings_radius = 1,
  sightings_opacity = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{eez_show}{Boolean; display EEZ boundaries?}

\item{eez_color}{Character string; color of EEZ boundary lines, if \code{eez_show} is \code{TRUE}.}

\item{eez_weight}{Single positive numeric; weight (thickness) of EEZ boundary lines, if \code{eez_show} is \code{TRUE}.}

\item{eez_opacity}{Single numeric between 0 and 1; opacity of EEZ boundary lines, if \code{eez_show} is \code{TRUE}.}

\item{strata_show}{Boolean; display geostratum boundaries?}

\item{strata_color}{Character string; color of geostratum boundary lines, if \code{strata_show} is \code{TRUE}.}

\item{strata_weight}{Single positive numeric; weight (thickness) of geostratum boundary lines, if \code{strata_show} is \code{TRUE}.}

\item{strata_opacity}{Single numeric between 0 and 1; opacity of geostratum boundary lines, if \code{strata_show} is \code{TRUE}.}

\item{effort_show}{Boolean; display survey tracks? Default is \code{FALSE}, because rendering the tracklines can take a while.
If \code{verbose} is \code{TRUE}, a progress bar will be printed to the console as the survey tracklines are rendered.}

\item{effort_resolution}{Single numeric integer equal to or larger than 1, indicating the data resolution of survey tracklines, if \code{effort_show} is \code{TRUE}.
A value of \code{1} means full resolution (no data rows are removed), and can take minutes to render.
Higher values mean lower resolution; in practice, the number provided represents the number of data lines skipped.
For example, a value of \code{3} means that only every fourth data line is used to build tracks.
For your initial call, we recommend the default (\code{15}).}

\item{effort_color}{Single character string; color of survey tracklines, if \code{effort_show} is \code{TRUE}.}

\item{effort_weight}{Single positive numeric; weight (thickness) of survey tracklines, if \code{effort_show} is \code{TRUE}.}

\item{effort_opacity}{Single numeric between 0 and 1; opacity of survey tracklines, if \code{effort_show} is \code{TRUE}.}

\item{sightings_show}{Boolean; display sightings?}

\item{sightings_color}{A single character string, indicating the color for sighting markers.
\strong{Note:} if you enter \code{"color code"}, the sightings will be color-coded by \code{leaflet} according to species code,
and a legend will be provided; this argument is ignored if \code{sightings_show} is \code{FALSE}.}

\item{sightings_radius}{A single postive numeric, indicating the size of the sightings markers; this argument is ignored if \code{sightings_show} is \code{FALSE}.}

\item{sightings_opacity}{A single numeric between 0 and 1, indicating the opacity of sightings markers; this argument is ignored if \code{sightings_show} is \code{FALSE}.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.
(The only updates occur in the preparation of effort tracks, if \code{effort_show} is \code{TRUE})}
}
\value{
An interactive \code{leaflet} map.
}
\description{
Produced an interactive \code{leaflet} map of your \code{Wincruz} survey data.
}
