% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_effort.R
\name{map_effort}
\alias{map_effort}
\title{Add tracklines to a survey map}
\usage{
map_effort(
  m,
  cruz,
  cohort = 1,
  use_type = c(TRUE),
  effort_color = "steelblue3",
  effort_stroke = 1.5,
  effort_linetype = 1,
  max_km_gap = NULL,
  verbose = FALSE
)
}
\arguments{
\item{m}{Your \code{tmap} map object (produced from \code{map_base()} or any of the other \code{LTabundR} mapping functions).}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{use_type}{Segment tracks to display, based on the \code{use} column within \verb{cruz$cohorts$<cohort>$<analysis>$segments}, and provided
as a boolean vector (accepted values are \code{TRUE} and/or \code{FALSE}).}

\item{effort_color}{Color of tracklines.
Before using this value, the function will first look for a \code{col} column in
\verb{cruz$cohorts$<cohort>$segments}.
If that is not found, the color provided here will be used.}

\item{effort_stroke}{Line width of tracklines.
Before using this value, the function will first look for a \code{lwd} column in
\verb{cruz$cohorts$<cohort>$segments}.
If that is not found, the stroke value provided here will be used.}

\item{effort_linetype}{Line type (classic graphical parameter \code{lty} values).
Before using this value, the function will first look for a \code{lty} column in
\verb{cruz$cohorts$<cohort>$segments}.
If that is not found, the value provided here will be used.}

\item{max_km_gap}{Maximum allowable gap between events, before breaking a continuous printed line.}

\item{verbose}{Boolean indicating whether or not to print status updates to the Console.}
}
\value{
A \code{tmap} map with survey tracklines shown (the map is built using the package \code{tmap}), which will print as a plot if not saved as an object.
If saved as an object, other features can be added to the map (see \code{map_strata()}, and \code{map_sightings()},
or other \code{tmap} functions).
}
\description{
Add tracklines to a survey map
}
