% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sightings.R
\name{map_sightings}
\alias{map_sightings}
\title{Add sightings to a survey map}
\usage{
map_sightings(
  m,
  cruz,
  cohort = 1,
  include_species = NULL,
  color_base = "firebrick",
  color_code = FALSE,
  shape_base = 21,
  shape_code = FALSE,
  size_base = 0.5,
  size_code = FALSE
)
}
\arguments{
\item{m}{Your \code{tmap} map object (produced from \code{map_base()} or any of the other \code{LTabundR} mapping functions).}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{include_species}{A character vector of species codes (see \code{data(species_codes)}), if you wish to only map sightings of certain species.}

\item{color_base}{Default color to use.}

\item{color_code}{Boolean; if \code{TRUE}, the underlying \code{tmap} functions will color-code sightings according to their built-in palettes.}

\item{shape_base}{Default point shape to use.}

\item{shape_code}{Boolean; if \code{TRUE}, the underlying \code{tmap} functions will shape-code sightings according to the \code{included} column.}

\item{size_base}{Default point size to use.}

\item{size_code}{Boolen; if \code{TRUE}, the underlying \code{tmap} functions will size-code sightings according to the \code{best} column.}
}
\value{
A \code{tmap} map with sightings shown as points (the map is built using the package \code{tmap}), which will print as a plot if not saved as an object.
If saved as an object, other features can be added to the map (see \code{map_effort()}, \code{map_strata()},
or other \code{tmap} functions)
}
\description{
Add sightings to a survey map
}
