% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_strata.R
\name{map_strata}
\alias{map_strata}
\title{Add geostratum polygons to a survey map}
\usage{
map_strata(
  m,
  settings,
  study_area_color = "forestgreen",
  study_area_stroke = 1,
  strata_color = "black",
  strata_stroke = 1,
  region = "cnp"
)
}
\arguments{
\item{m}{Your \code{tmap} map object (produced from \code{map_base()} or any of the other \code{LTabundR} mapping functions).}

\item{settings}{A \code{settings} object, as produced by \code{LTabundR::load_settings()}.}

\item{study_area_color}{Color of study area boundary line.}

\item{study_area_stroke}{Width of study area boundary line.}

\item{strata_color}{Color of geostratum boundary line(s).}

\item{strata_stroke}{Width of geostratum boundary line(s).}

\item{region}{Name of the survey region you wish to map.
Options currently supported:
\code{"cnp"} (the default; Central North Pacific), \code{"ccs"} (California Current System),
and \code{"etp"} (Eastern Tropical Pacific). This argument will define how coordinates are re-projected to handle the international date line.}
}
\value{
A \code{tmap} map with geostratum boundaries shown (the map is built using the package \code{tmap}), which will print as a plot if not saved as an object.
If saved as an object, other features can be added to the map (see \code{map_effort()}, \code{map_sightings()},
or other \code{tmap} functions).
}
\description{
Add geostratum polygons to a survey map
}
