% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_group_affiliates.R
\name{mixed_group_affiliates}
\alias{mixed_group_affiliates}
\title{Find affiliated species from mixed-species groups}
\usage{
mixed_group_affiliates(cruz, cohort = 1, spp)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose sightings you would like to recalibrate,
provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{spp}{Character vector of species codes. Groups containing these species
will be found, and all other species present in their mixed-species groups will
be returned as "affiliates".}
}
\value{
A character vector of species codes for all species affiliated with \code{spp} in
mixed-species groups, provided in increasing order of frequency.
}
\description{
Find affiliated species from mixed-species groups
}
