% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa_daily_1986_2020}
\alias{noaa_daily_1986_2020}
\title{Cruise data processed for all NOAA SWFSC and PIFSC surveys, 1986 - 2020 (daily segment lengths)}
\format{
An object of class \code{list} of length 3.
}
\usage{
noaa_daily_1986_2020
}
\description{
Cruise data processed for all NOAA SWFSC and PIFSC surveys, 1986 - 2020 (daily segment lengths)
}
\details{
This dataset was processed using the following code:\preformatted{data(group_size_coefficients)

survey <- load_survey_settings(
out_handling = 'stratum',
max_row_interval = Inf,
segment_method = "day",
group_size_coefficients = group_size_coefficients)

data(strata_ccs)
data(strata_cnp)
strata <- c(strata_ccs, strata_cnp)
strata$NEPAC <- data.frame(Lat = c(-20, -20, 65, 65), Lon = c(-230, -70, -70, -200))

all_species <- load_cohort_settings(
id = "all",
species = NULL,
probable_species = FALSE,
sighting_method = 0,
cue_range = 0:7,
school_size_range = c(0, 10000),
school_size_calibrate = TRUE,
calibration_floor = 0,
use_low_if_na = TRUE,
io_sightings = 0,
geometric_mean_group = TRUE,
truncation_km = 7.5,
beaufort_range = 0:6,
abeam_sightings = TRUE,
strata_overlap_handling = c("smallest"),
distance_types = c('S','F','N'),
distance_modes = c('P','C'),
distance_on_off = TRUE
)

settings <- load_settings(strata = strata,
                       study_area = NULL,
                       survey = survey,
                       cohorts = list(all_species))

das_file = c('../test_code/data/swfsc_1986_2020.das')
swfsc <- process_surveys(das_file,
                      settings = settings)

das_file = '../test_code/CNP/CenPac1986-2020_Final_alb.das'
pifsc <- process_surveys(das_file,
                      settings = settings)

cruzes <- list(swfsc, pifsc)
noaa_daily_1986_2020 <- cruz_combine(cruzes)
}
}
\keyword{datasets}
