% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_points_to_line.R
\name{points_to_line}
\alias{points_to_line}
\title{Convert coordinates to a \code{sf} line}
\usage{
points_to_line(data, long, lat, id_field = NULL, sort_field = NULL)
}
\arguments{
\item{data}{The dataframe containing columns with coordinates.}

\item{long}{A character vector with the name of the column with longitude coordinates.}

\item{lat}{A character vector with the name of the column with latitude coordinates.}

\item{id_field}{If the dataframe contains data for multiple lines, specify the
name of the column (as a character vector) that indicates which line each point belongs to.}

\item{sort_field}{If the dataframe is not arranged in the sequence you wish, specify the
name of the column (as a character vector) you want to arrange by before creating the line.}
}
\value{
A \code{sp::SpatialLines} object created from \code{sp::spRbind()}.
}
\description{
This is an internal function typically not called by a user directly.
}
