% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bootstrap_datasets.R
\name{prep_bootstrap_datasets}
\alias{prep_bootstrap_datasets}
\title{Prepare bootstrapped versions of datasets}
\usage{
prep_bootstrap_datasets(segments, sightings, segment_picks = NULL)
}
\arguments{
\item{segments}{Effort segments dataframe, drawn from a \code{cruz} object (e.g., \code{cruz$cohorts[[1]]$segments}),
already filtered to contain the effort you wish to use to fit the detection function.}

\item{sightings}{Sightings dataframe, drawn from a \code{cruz} object (e.g., \code{cruz$cohorts[[1]]$sightings}),
already filtered to contain the sightings you wish to use to fit the detection function.}

\item{segment_picks}{If segment IDs have been previously selected for this dataset, provide them here.}
}
\value{
A list with resampled data: \code{segments}, \code{sightings}, and \code{segment_picks}
(a numeric vector of segment IDs that have been re-sampled).
}
\description{
An interior function, typically not called by analysts.
This function returns segment and sighting data according to resampled segment IDs,
which serve as part of the bootstrap variance estimation process.
The function includes an option to specify a selection of segment IDs, which allows
the same bootstrap segments to be selected for the abundance estimation stage as was
used in the detection function model fitting stage.
}
