% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_km.R
\name{process_km}
\alias{process_km}
\title{Calculate distances between rows in a DAS file}
\usage{
process_km(
  das,
  min_interval = 5,
  max_interval = 3600,
  replacement_interval = 3600,
  max_km_gap = 100,
  km_filler = 1,
  speed_filler = 10 * 1.852,
  debug_mode = TRUE
)
}
\arguments{
\item{das}{A dataframe of \code{DAS} survey data.}

\item{min_interval}{Minimum seconds between rows before distance will be calculated.
Default is 30 seconds; if the interval is less than this value, the distance will be assumed to be 0 km.}

\item{max_interval}{The maximum time interval, in seconds, between rows before assuming that
there has been a break in survey data logging. This time interval will be replaced with the value specified in \code{replacement_interval}.}

\item{replacement_interval}{The time interval, in second, to use as a replacement
when the time interval between rows exceeds \code{max_interval}. The default is 900 seconds.
or the \code{max_interval} specified, whichever is smaller.}

\item{max_km_gap}{Another way of avoiding long gaps in data; this is the maximum distance gap, in km, allowed between rows of data.
This constraint is applied \emph{after} the \code{interval} constraints above.}

\item{km_filler}{When valid speed and position information is not available (e.g., the given distance exceeds \code{max_km_gap}), this value (in km) will be used as an estimate of the
distance in between consecutive rows of data.}

\item{speed_filler}{When speed is not available in the data, this value (in kmh) will be used as a filler in order to estimate the
distance between consecutive rows of data based on timestamp differences (when lat/long coordinates are not available).}

\item{debug_mode}{Boolean, with default \code{FALSE}, indicating whether details should be printed to the Console that facilitate debugging.}
}
\value{
A numeric vector, the same length as the number of rows in \code{das}, of distances (in km).
The final element of this vector will be 0.
}
\description{
This function is the near-equivalent of the \code{DISTRAV} subroutine in \code{ABUND9}.
This is an internal function typically not called by a user directly.
It is called in \code{format_das()}.
}
