% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polygon.R
\name{process_polygon}
\alias{process_polygon}
\title{Format a CSV into a polygon for mapping}
\usage{
process_polygon(polygon_dataframe, remove_land = TRUE, toplot = FALSE)
}
\arguments{
\item{polygon_dataframe}{A dataframe in which the first two columns are \code{Lon} and \code{Lat},
providing coordinates in decimal degrees in which South and West coordinates are negative.
It is acceptable if vertices in the eastern hemisphere are described using negative longitudes below -180, e.g., -185.
(The function will correct these to proper decimal degrees, e.g., -185 will become 175.)
Other columns are allowed, but the first two need to be \code{Lon} and \code{Lat} (in that order).}

\item{remove_land}{If \code{TRUE}, the area of land occurring within the polygon will
be subtracted from the polygon's area, so that the reported area represents aquatic habitat only.}

\item{toplot}{If \code{TRUE} (not the default), a map of the polygon will be plotted,
and -- if \code{remove_land} is \code{TRUE} -- any land inside will be plotted in grey.}
}
\value{
A list with four named slots:
\enumerate{
\item \code{coords} The input dataframe, with longitudes corrected to proper negative decimal degrees, e.g., longitudes supplied as -185 will become 175.
\item \code{polygon}  The polygon created by \code{sf::st_polygon()}
\item \code{sf}  The \code{sf} version of the polygon that will be passed to mapping functions.
\item \code{sf_land} The \code{sf} version of the land polygons included within the polygon, if any.
\item \code{km2}  The area of the polygon, in square kilometers.
}
}
\description{
This is an internal function typically not called by a user directly.
It takes a dataframe of polygon coordinates and returns a list with the polygon formatted in various ways,
as well as a calculation of the area of the polygon \emph{with land area removed} (optionally).
}
