% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_strata.R
\name{process_strata}
\alias{process_strata}
\title{Process geo-strata}
\usage{
process_strata(das, settings, verbose = TRUE)
}
\arguments{
\item{das}{A \code{data.frame} of a \code{DAS} survey data file, created by \code{load_das()}.}

\item{settings}{A \code{settings} object, created by \code{load_settings()}.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A list with two named slots: [\link{1}] \code{settings} and [\link{2}] \code{das}.
The \code{das} slot has the dataframe of your survey data.
For each stratum, a column named \verb{stratum_<StratumName>} is added to the \code{das} dataframe;
each row in this column is \code{TRUE} (included) or \code{FALSE}.
}
\description{
Determine in/out status of \code{DAS} events for each geo-strata provided by \code{settings}.
This function loops through each stratum dataframe you have provided it in \code{settings$strata},
formats the stratum, and asks whether each DAS row occurs within it.
\cr \cr
This is an internal function typically not called by a user directly.
It is the second subroutine called within \code{process_surveys()}, after \code{load_das()} and before \code{format_das()}.
}
