% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_subgroups.R
\name{process_subgroups}
\alias{process_subgroups}
\title{Analyze subgroup detections for false killer whales}
\usage{
process_subgroups(cruz, edits = NULL, verbose = TRUE)
}
\arguments{
\item{cruz}{A \code{cruz} object passed from \code{process_surveys()}.}

\item{edits}{An optional input: a \code{list} indicating manual adjustments to the subgroup \code{events} data.
This \code{list} can contain elements produced by the \code{LTabundR} functions \code{subgroup_populations()} or \code{subgroup_edits()},
or simply \code{list}(s) that the user prepares themselves.

Each \code{list} element needs to be either a \code{data.frame} or a \code{list}
with the following names: \code{edit}, \code{cohort}, \code{crusie}, and \code{sgid}. The remaining names depend on the type of \code{edit}
(see the documentation for the \code{subgroup_edits()} function for details). Edits will be applied in the order in which they are supplied.

If this input is supplied, the \code{process_subgroups()} function will first process the data using default operations,
then redact the \code{events} data according to this input. The recommended workflow here is:
(1) process your survey data without supplying \code{edits};
(2) review the \code{events} data using the \code{subgroup_explorer()} function;
(3) if revisions are needed, stage those edits using \code{subgroup_populations()} function, \code{subgroup_edits()} function, or manually coded lists.
(4) collate those staged edits into a list.
(5) re-run \code{process_subgroups()}, this time supplying the \code{edits} input.

See the vignette for details.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A finalized \code{cruz} object.
If subgroups are found, a \code{subgroups} slot is added to the analysis list for a cohort.
This \code{subgroups} slot holds a list with three dataframes:
\enumerate{
\item \code{events}, in which each row is a school size estimate for a single subgroup during a single phase -- 1 or 2 -- within a single sighting;
\item \code{subgroups}, in which each row is a single phase for a single subgroup, with all school size estimates averaged together (both arithmetically and geometrically);
\item \code{sightings}, in which each row is a school size estimate for a single phase for a single sighting, with all subgroup school sizes summed together.
}
Phase is determined simply according to the \code{OnEffort} column. If it is \code{TRUE}, \code{Phase} is 1; if not, \code{Phase} is 2.
}
\description{
Subgroups are found based on \code{DAS} event code "\code{G}" and associated events.
Subgroup sizes are estimated for each phase (based upon the false killer whale protocol),
and both arithmetic and geometric means are provided.
\cr \cr
This is an internal function typically not called by a user directly.
It is the final processing subroutine called within \code{process_surveys()}, after the subroutine \code{process_sightings()}.
}
