% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polygon.R
\name{read_polygon}
\alias{read_polygon}
\title{Read in a DAT file with polygon(s)}
\usage{
read_polygon(dat_file)
}
\arguments{
\item{dat_file}{Filepath to a \code{.DAT} file containing coordinates for one or more polygons.}
}
\value{
A named list:
\enumerate{
\item \code{polygons} A dataframe summarizing the polygons contained in the \code{.DAT} file.
It has four columns: \code{name} (polygon name), \code{area} (polygon area, if provided),
\code{name_i} (the index of the polygon in the file), and \code{area_i} (the index of the area provided in the file).
\item \code{coordinates} A list with a slot for each polygon contained within the \code{.DAT} file. Each polygon is a
dataframe with two columns: \code{Lon} and \code{Lat}, with coordinates formatted as provided by the \code{.DAT} file.
\item \code{dat} The raw \code{DAT} file, as it was read by \code{R}, returned as a dataframe.
}
}
\description{
This is an internal function typically not called by a user directly.
This function reads in a \code{.DAT} file with coordinates for one or more polygons,
returning a list.
}
