% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_das.R
\name{region_das}
\alias{region_das}
\title{Find DAS rows occurring within a region / regions}
\usage{
region_das(das, regions)
}
\arguments{
\item{das}{The \code{DAS} table from a \code{cruz} object.}

\item{regions}{A character vector of geostrata whose segments you want.}
}
\value{
A vector of \code{line_num} values whose data occur within any of \code{regions},
which can be used to filter segments and sightings.
}
\description{
This is an internal function, typically not called by the user.
}
