% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_segments.R
\name{region_segments}
\alias{region_segments}
\title{Find segments occurring within a region / regions}
\usage{
region_segments(das, regions)
}
\arguments{
\item{das}{The \code{DAS} table from a \code{cruz} object.}

\item{regions}{A character vector of geostrata whose segments you want.}
}
\value{
A vector of segment IDs whose data occur within \code{regions},
which can be used to filter segments and sightings by the \code{seg_id} column.
}
\description{
This is an internal function, typically not called by the user.
}
