% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize.R
\name{segmentize}
\alias{segmentize}
\title{Segmentize survey effort}
\usage{
segmentize(
  cruz,
  segment_method = NULL,
  segment_target_km = NULL,
  segment_max_interval = NULL,
  segment_remainder_handling = NULL,
  beaufort_range = NULL,
  distance_types = NULL,
  distance_modes = NULL,
  distance_on_off = NULL,
  to_plot = TRUE,
  debug_mode = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cruz}{A nascent \code{cruz} object, which is a list produced by \code{format_das()},
or a fully processed \code{cruz} object if you want to experiment with segmentizing alternatives.}

\item{segment_method}{This and the remainder of inputs allow you to override
the settings contained within the \code{cruz} object. Leaving them as \code{NULL}
means the original settinsg will be used.
For \code{segment_method}, the two method options are \code{"day"} --
all effort within the same Cruise-Stratum-Year-Effort scenario (i.e., an effort bloc)
will be binned into segments by calendar date -- and \code{"equallength"} --
effort within each unique effort bloc
will be divided into segments of approximately equal length.}

\item{segment_target_km}{If segmentizing by \code{"equallength"},
this field allows you to specify what that target length is, in km.}

\item{segment_max_interval}{If segmentizing by \code{"equallength"},
this setting allows you to specify the time gaps in effort
that are allowed to be contained within a single segment.
For example, if your goal is a few large segments of equal length
(e.g., 150-km segments, for bootstrap estimation of density variance),
you are probably willing for discrete periods of effort to be concatenated into a single segment,
even if the gaps between effort are as large as 1 or 2 days,
in which case you would set \code{segment_max_interval} to 24 or 48 (hours), respectively.
However, if your goal is many smaller segments (e.g., 5-km segments, for habitat modeling),
you want to ensure that effort is contiguous so that segment locations
can be accurately related to environmental variables,
in which case you would set \code{segment_max_interval} to be very small (e.g., .2 hours, or 12 minutes).
Setting this interval to a small number, such as 0.2, also allows
the segmentizing function to overlook momentary breaks in effort,
such as when an unofficial observer logs a sighting.}

\item{segment_remainder_handling}{If segmentizing by \code{"equallength"},
periods of effectively-contiguous effort (as specified by \code{segment_max_interval})
are unlikely to be perfectly divisible by your \code{segment_target_km};
there is going to be a remainder. You can handle this remainder in three ways:
(1) \code{"disperse"} allows the function to adjust \code{segment_target_km} so that
there is in fact no remainder, effectively dispersing the remainder evenly
across all segments within that period of contiguous effort;
(2) \code{"append"} asks the function to append the remainder to a randomly selected segment,
such that most segments are the target length with the exception of one longer one;
or (3) \code{"segment"} asks the function to simply place the remainder in its own segment,
placed randomly within the period of contiguous effort.
This setting also has a second layer of versatility,
because it can accept a one- or two-element character vector.
If a two-element vector is provided (e.g., \code{c("append","segment")}),
the first element will be used in the event that the remainder is less than or equal to
half your \code{segment_target_km}; if the remainder is more than half that target length,
the second element will be used. This feature allows for replication
of the segmentizing methods in Becker et al. (2010).}

\item{beaufort_range}{A numeric vector indicating the Beaufort sea states (0 - 7) to be accepted within usable segments.}

\item{distance_types}{A character vector of the effort types that will be included
in detection function estimation,
and therefore considered in effort segmentizing.
Accepted values are \code{"S"} (systematic/standard effort), \code{"F"} (fine-scale effort),
and \code{"N"} (non-systematic/non-standard effort, in which systematic protocols
are being used but effort is not occurring along design-based transect routes).}

\item{distance_modes}{The effort modes that will be included in
detection function estimation,
and therefore considered in effort segmentizing.
Accepted values are \code{"P"} (passing) and \code{"C"} (closing)}

\item{distance_on_off}{The value(s) of \code{OnEffort}
(On Effort is \code{TRUE}, Off Effort is \code{FALSE}) that will be included in
detection function estimation,
and therefore considered in effort segmentizing.}

\item{to_plot}{Boolean, with default \code{FALSE}, indicating whether or not histograms showing segment lengths should be produced.}

\item{debug_mode}{Boolean, with default \code{FALSE}, indicating whether details should be printed to the Console that facilitate debugging.}

\item{verbose}{Boolean, with default \code{FALSE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A modified \code{cruz} object, in which each cohort slot contains a list with the following slots:
\code{das} (with two new columns, \code{seg_id}, indicating the segment corresponding to each \code{DAS} row;
and \code{use}, indicating whether or not (\code{TRUE} or \code{FALSE}) the segment is to be included in the analysis.)
and \code{segments} (a dataframe summarizing each segment).
}
\description{
This function chops \code{DAS} effort into discrete sections for the purposes of estimating the variance of the abundance estimate.
\cr \cr
This is an internal function typically not called by a user directly.
It is the fourth subroutine called within \code{process_surveys()}, after \code{format_das()} and before \code{process_sightings()}.
However, arguments are provided that allow you to call this function directly and override settings so that you can
explore how segmentizing works.
}
