% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_segmentize_core.R
\name{segmentize_core}
\alias{segmentize_core}
\title{Core utility function for segmentizing effort}
\usage{
segmentize_core(
  dasspliti,
  segment_method,
  segment_target_km,
  segment_max_interval,
  segment_remainder_handling,
  debug_mode,
  verbose
)
}
\arguments{
\item{dasspliti}{An "effort bloc" of survey data corresponding to a single effort scenario
(year, cruise number, stratum, effort type, effort mode, etc.).}

\item{segment_method}{See documentation for \code{segmentize()}.}

\item{segment_target_km}{See documentation for \code{segmentize()}.}

\item{segment_max_interval}{See documentation for \code{segmentize()}.}

\item{segment_remainder_handling}{See documentation for \code{segmentize()}.}

\item{debug_mode}{Boolean indicating whether or not details should be printed to the Console that facilitate debugging.}

\item{verbose}{Boolean, with default \code{FALSE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A list in which each slot is a chunk of survey effort corresponding to a single segment.
}
\description{
This is an internal function typically not called by a user directly.
It is called within \code{segmentize_wrapper()}, which itself is a subroutine
called within \code{segmentize()}.
This function segmentizes the data within a single "effort bloc", returning a list of segments.
}
