% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_segmentize_summarize.R
\name{segmentize_summarize}
\alias{segmentize_summarize}
\title{Utility function to summarize segments}
\usage{
segmentize_summarize(
  segments,
  das,
  debug_mode = FALSE,
  to_plot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{segments}{A list in which each slot has a chunk of \code{DAS} survey data corresponding to a single effort segment.}

\item{das}{The \code{DAS} survey data.}

\item{debug_mode}{Boolean, with default \code{FALSE}, indicating whether details should be printed to the Console that facilitate debugging.}

\item{to_plot}{Boolean, with default \code{TRUE}, indicating whether or not histograms showing segment lengths should be produced.}

\item{verbose}{Boolean, with default \code{FALSE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A list with two slots: \code{segments} (summary dataframe, in which each row contains metadata for a single segment),
and \code{das} (modified \code{DAS} survey data, with new columns \code{segid} -- segment ID -- and \code{use} -- a logical indicating
whether the event will be included in analysis). This result is passed back to \code{segmentize()}.
}
\description{
This is an internal function typically not called by a user directly.
This function takes a list with segment data and produces a dataframe
summarizing each segment.
}
