% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_segmentize_wrapper.R
\name{segmentize_wrapper}
\alias{segmentize_wrapper}
\title{Segmentizer wrapper to prevent redundant code}
\usage{
segmentize_wrapper(
  das,
  segment_method,
  segment_target_km,
  segment_max_interval,
  segment_remainder_handling,
  beaufort_range,
  types,
  modes,
  on_off,
  debug_mode,
  verbose
)
}
\arguments{
\item{das}{The \code{DAS} survey data.}

\item{segment_method}{See documentation for \code{segmentize()}.}

\item{segment_target_km}{See documentation for \code{segmentize()}.}

\item{segment_max_interval}{See documentation for \code{segmentize()}.}

\item{segment_remainder_handling}{See documentation for \code{segmentize()}.}

\item{types}{See documentation for \code{segmentize()}.}

\item{modes}{See documentation for \code{segmentize()}.}

\item{on_off}{See documentation for \code{segmentize()}.}

\item{debug_mode}{Boolean, with default \code{FALSE}, indicating whether details should be printed to the Console that facilitate debugging.}

\item{verbose}{Boolean indicating whether or not updates should be printed to the Console.}
}
\value{
A list in which each slot is a chunk of survey effort corresponding to a single segment.
This result is passed back to \code{segmentize()}.
}
\description{
This is an internal function typically not called by a user directly.
It is called within \code{segmentize()}.  This function parses \code{DAS} survey data
into "effort blocs", in which effort conditions are the same.
It then calls a subroutine, \code{segmentize_core()} to segmentize the data within
each effort bloc.
}
