% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ships}
\alias{ships}
\title{Ship codes}
\format{
An object of class \code{data.frame} with 77 rows and 3 columns.
}
\usage{
ships
}
\description{
Ship codes
}
\details{
A \code{data.frame} matching the ship name initials to the NOAA-NFMS cruise number.
}
\keyword{datasets}
