% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Species codes used in WinCRUZ}
\format{
A \code{data.frame} with 164 rows and 7 variables:
\describe{
\item{code}{Code used in WinCRUZ}
\item{short_name}{Abbreviated name, in all CAPS}
\item{scientific_name}{Latin Genus species}
\item{common_name1}{Most common name}
\item{common_name2}{Alternative common name}
\item{comman_name3}{Alternative common name}
\item{common_name4}{Alternative common name}
}
}
\usage{
species_codes
}
\description{
Species codes used in WinCRUZ
}
\details{
A \code{data.frame} containing species codes and associated identifiers.
}
\keyword{datasets}
