% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_translator.R
\name{species_translator}
\alias{species_translator}
\title{Species name/code search engine}
\usage{
species_translator(
  id,
  codes = NULL,
  match_by_code = TRUE,
  match_by_short_name = TRUE,
  match_by_latin = TRUE,
  match_by_common = TRUE
)
}
\arguments{
\item{id}{Your search query term; can be a character or numeric. The function will
return partial matches and is case insensitive.}

\item{codes}{A \code{data.frame} of species codes, with the same column structure
as the \code{LTabundR} dataset \code{data(species_codes)}. If not provided by the user, that
built-in dataset will be provided.}

\item{match_by_code}{The remaining inputs are a way to toggle on/off certain
methods for searching for matches. This input will try to match your search query
to species codes, e.g., \code{"033"} for the false killer whale.}

\item{match_by_short_name}{Try matching search query to "short name" species codes,
e.g., \code{"MESOP_PERU"} for the pygmy beaked whale.}

\item{match_by_latin}{Try matching search query to scientific name.}

\item{match_by_common}{Try matching search query to common name.}
}
\value{
A \code{data.frame} with all viable patches to your search query.
}
\description{
Retrieve common and scientific names for a species based on their code number or short name,
or vice versa: retrieve the species code based on common or scientific name.
}
\examples{
species_translator('046') # match by species code
species_translator('46') # partial matches work
species_translator(46) # numeric codes work
species_translator('RM_WHA') # partial short names work
species_translator('Physeter') # latin names work
species_translator('orca') # common names work + partial matches + case insensitive
species_translator('orca', match_by_latin=FALSE) # toggle match modes

}
