% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata_ccs}
\alias{strata_ccs}
\title{Strata CCS}
\format{
An object of class \code{list} of length 5.
}
\usage{
strata_ccs
}
\description{
Strata CCS
}
\details{
Typical geostrata used in the California Current System region.
}
\keyword{datasets}
