% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata_cnp}
\alias{strata_cnp}
\title{Strata CNP}
\format{
An object of class \code{list} of length 11.
}
\usage{
strata_cnp
}
\description{
Strata CNP
}
\details{
Typical geostrata used in the Central North Pacific region.
}
\keyword{datasets}
