% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strata_etp}
\alias{strata_etp}
\title{Strata ETP}
\format{
An object of class \code{list} of length 70.
}
\usage{
strata_etp
}
\description{
Strata ETP
}
\details{
Typical geostrata used in the Eastern Tropical Pacific region.
}
\keyword{datasets}
