% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_explore.R
\name{strata_explore}
\alias{strata_explore}
\title{Explore the package's strata polygon gallery}
\usage{
strata_explore(
  region,
  start_at = 1,
  strata_color = "darkblue",
  begin_hidden = FALSE
)
}
\arguments{
\item{region}{Name of the survey region of interest. Options currently supported:
\code{"cnp"} (the default; Central North Pacific), \code{"ccs"} (California Current System),
and \code{"etp"} (Eastern Tropical Pacific).#'}

\item{start_at}{A number indicating the geostratum index you wish to begin at, if you know you aren't interested in those early on in the list. For example, the ETP has over 70 built-in geostrata available!}

\item{strata_color}{Color of geostratum.}

\item{begin_hidden}{If \code{TRUE} (not the default), the strata will begin hidden, i.e.,
unchecked, so that you have to opt to visualize each stratum individually.
This can be handy for regions with many geostrata to choose from, such as the
ETP, which has 70 geostrata!}
}
\value{
An interactive \code{leaflet} map.
The name of the geostratum, and the numeric identifier for it, are printed
in a legend as well as in a pop-up box that appears when you hover over a stratum.
Use this identifier as an argument in the companion function, \code{strata_select()}, to access the datasets of coordinates
underlying the polygons displayed.
}
\description{
The \code{LTabundR} package comes with several built-in datasets of geographic strata
that are commonly used in NOAA/NMFS surveys. The functions \code{strata_explore()}
and \code{strata_select()} were developed to help you explore those built-in options.
\cr \cr
\code{strata_explore()} launches a \code{leaflet} map displaying the built-in
geostrata datasets available for your use in the common NOAA/NMFS survey regions.
Each stratum displayed as a separate layer that can be hidden or hovered over to get details.
}
