% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_select.R
\name{strata_select}
\alias{strata_select}
\title{Select strata polygons for standard survey regions}
\usage{
strata_select(selections, region)
}
\arguments{
\item{selections}{Numeric index of the geostratum you wish to "grab" from a
list of geostrata. This index is printed in the legend produced by \code{strata_explore()}.}

\item{region}{Name of the survey region pertaining to your selection (the same region you passed to \code{strata_explore()}).
Options currently supported:
\code{"cnp"} (the default; Central North Pacific), \code{"ccs"} (California Current System),
and \code{"etp"} (Eastern Tropical Pacific).}
}
\value{
A \code{data.frame} of coordinates for the boundary of a geostratum. You can
pass this object to the \code{strata} argument for \code{load_settings()}.
}
\description{
The \code{LTabundR} package comes with several built-in datasets of geographic strata
that are commonly used in NOAA/NMFS surveys. The functions \code{strata_explore()}
and \code{strata_select()} were developed to help you explore those built-in options.
\cr \cr
Those built-in datasets come in the form of a list, in which each slot is a
separate geostratum. \code{strata_select()} allows you to "grab" the underlying data (i.e., the \code{data.frame}
of Lat/Long coordinates) for one of the geostrata displayed by \code{strata_explore()}.
}
