% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_edits.R
\name{subgroup_edits}
\alias{subgroup_edits}
\title{Stage edits to subgroups (e.g., phase and population assignments)}
\usage{
subgroup_edits(
  cohort,
  sgid,
  exclude = NULL,
  phase = NULL,
  population = NULL,
  pop_prob = NULL
)
}
\arguments{
\item{cohort}{The cohort whose data will be edited,
provided as the number or name of the slot in \code{cruz$cohorts} to be referenced.}

\item{sgid}{A character vector of subgroup IDs whose data will be edited.}

\item{exclude}{A Boolean vector, with length of either 1 or the same as \code{sgid},
indicating which \code{sgid}'s to erase: \code{TRUE} means exclude, \code{FALSE} means keep.}

\item{phase}{A numeric vector, with length of either 1 or the same as \code{sgid},
indicating the new phase to assign the \code{sgid}'s.
If you wish to remove a phase assignment for a \code{sgid}, use \code{NA}.}

\item{population}{A character vector, with length of either 1 or the same as \code{sgid},
indicating the new population to assign the \code{sgid}'s. If you want a \code{sgid} to be eligible for multiple
populations, separate the population names with a semicolon.
Both this input and \code{pop_prob} needs to not be \code{NULL} in order for this input to be recognized.}

\item{pop_prob}{A numeric vector, with length of either 1 or the same as \code{sgid}
and values ranging between 0 and 1,
indicating the probability of each population assignmend for the \code{sgid}'s.
If you want a \code{sgid} to be eligible for multiple
populations, separate the probabilities by a semicolon.
Both this input and \code{population} needs to not be \code{NULL} in order for this input to be recognized.}
}
\value{
A \code{list} of staged edits that will be accepted by the \code{edits} input in \code{process_subgroups()}.
}
\description{
Stage edits to subgroups (e.g., phase and population assignments)
}
