% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_populations.R
\name{subgroup_populations}
\alias{subgroup_populations}
\title{Assign subgroups to populations based on polygons}
\usage{
subgroup_populations(
  populations,
  cruz,
  cohort = "pseudorca",
  default_pop = "pelagic",
  verbose = TRUE
)
}
\arguments{
\item{populations}{A named list in which each slot is a \code{data.frame} of coordinates for a geostratum polygon.
Each \code{data.frame} must have \code{Lon} and \code{Lat} as the first two columns,
providing coordinates in decimal degrees in which West and South coordinates are negative.
It is acceptable if vertices in the eastern hemisphere are described using negative longitudes below -180, e.g., -185.
Other columns are allowed, but the first two need to be \code{Lon} and \code{Lat}.
The name of the slot holding the \code{data.frame} will be used as a reference name for the stratum.
Note that if coordinates in your data or in your collection of strata span
the International Date Line (IDL) such that some longitudes are positive
and some are negative, during data processing all longitudes will be coerced
to negative degrees West.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to use, provided as the number or name of the slot in \code{cruz$cohorts} to be referenced.}

\item{default_pop}{desc}

\item{verbose}{Boolean; print updates to console?}
}
\value{
A \code{data.frame} of population assignments for each subgroup event, formatted to be accepted as edits
as an input in \code{process_subgroups()}.
}
\description{
Assign subgroups to populations based on polygons
}
