% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_sightings.R
\name{subgroup_sightings}
\alias{subgroup_sightings}
\title{Summarize sightings (for subgroup species) based on subgroups}
\usage{
subgroup_sightings(subgroups)
}
\arguments{
\item{subgroups}{The output of \code{subgroup_subgroups()}.}
}
\value{
A \code{data.frame} with one row for each phase of each sighting.
}
\description{
This is an internal function, not typically called by the user.
It is called during the \code{process_subgroups()} routine within \code{process_surveys()},
and may also be called if the user manually edits subgroup phase assignments using the \code{subgroup_phases()} function.
It takes the result of the function \code{subgroup_subgroups()}.
}
