% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_subgroups.R
\name{subgroup_subgroups}
\alias{subgroup_subgroups}
\title{Summarize subgroups based on subgroup events}
\usage{
subgroup_subgroups(events)
}
\arguments{
\item{events}{}
}
\value{
A \code{data.frame} with one row for each subgroup in each phase of a sighting.
}
\description{
This is an internal function, not typically called by the user.
It is called during the \code{process_subgroups()} routine within \code{process_surveys()},
and may also be called if the user manually edits subgroup phase assignments using the \code{subgroup_phases()} function.
It takes the result of the function \code{subgroup_events()}.
}
