% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_distances.R
\name{summarize_distances}
\alias{summarize_distances}
\title{Summarize sighting distances in a table}
\usage{
summarize_distances(
  distances,
  distance_range = c(0, 10),
  distance_interval = 0.5
)
}
\arguments{
\item{distances}{Perpendicular distances to sightings, usually in km.}

\item{distance_range}{The range of distances for which to summarize detection distances.}

\item{distance_interval}{The interval of distances to summarize in the table.}
}
\value{
A \code{dataframe} with a row for each distance interval, summarizing the number of sightings, etc.
}
\description{
Summarize sighting distances in a table
}
