% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_effort.R
\name{summarize_effort}
\alias{summarize_effort}
\title{Summarize \code{Wincruz} survey effort}
\usage{
summarize_effort(cruz, cohort = 1)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to summarize, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}
}
\value{
A list with various summary tables:
\enumerate{
\item \code{total}: Grand total distance and days surveyed. Two columns: \code{km}, \code{days}.
\item \code{total_by_year}: Total distance and days surveyed for each year Three columns:
\code{year}, \code{km}, \code{days}.
\item \code{total_by_cruise}: Total distance and days surveyed for each cruise. Four columns:
\code{Cruise}, \code{year}, \code{km}, \code{days}.
\item \code{total_by_effort}: Total distance and days surveyed, grouped by segments that will be included in the analysis and those that won't.
Five columns: \code{Cruise}, \code{year}, \code{use} (used for analysis or not), \code{km}, and \code{days}.
\item \code{total_by_stratum}: Total distances and days surveyed within each stratum, again grouped by segments
that will be included in the analysis and those that won't. Six columns:
\code{Cruise}, \code{year}, \code{stratum}, \code{use}, \code{km}, \code{days}.
}
}
\description{
Inventory and summarize survey effort within a \code{LTabundR} \code{cruz} object in various ways.
}
