% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_species.R
\name{summarize_species}
\alias{summarize_species}
\title{Summarize sightings for a species (or several)}
\usage{
summarize_species(
  spp,
  cruz,
  cohort = 1,
  filter_to_regions = NULL,
  exclude_regions = NULL,
  distance_restrict = TRUE,
  distance_range = c(0, 10),
  distance_interval = 0.5
)
}
\arguments{
\item{spp}{A character vector with codes for the species you wish to summarize.}

\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to map, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}

\item{filter_to_regions}{Regions (geostrata) to filter the results down to. Use \code{cruz$strata} to see the options available to you.}

\item{exclude_regions}{Regions (geostrata) to exclude from the results. This can be helpful if, for example, you wish to
summarize sightings for a large region but not for regions nested within it, such as insular geostrata.}

\item{distance_restrict}{If \code{TRUE}, detection distances will only be shown for sightings that will be included in analysis (\code{included == TRUE}).}

\item{distance_range}{The range of distances for which to summarize detection distances.}

\item{distance_interval}{The interval of distances to summarize.}
}
\value{
A list with the following named slots:
\enumerate{
\item \code{species}: A dataframe with the codes, common names, and scientific name(s) for the specified species (as found in \code{data(species_codes)})
\item \code{n_total}: Total number of sightings (mixed-species are only counted once; this is true for all other slots here as well).
\item \code{n_analysis}: Number of sightings that will qualify for inclusion in the analysis (\code{included == TRUE}).
\item \code{school_size}: Dataframe with summary metrics of school size for each species.
\item \code{yearly_total}: Dataframe with counts and school sizes metrics for all sightings, parsed by year of sighting.
\item \code{yearly_analysis}: Dataframe with counts and school size metrics for sightings included in the analysis, parsed by year of sighting.
\item \code{regional_total}: Dataframe with counts and school sizes metrics for all sightings, parsed by regional geostratum.
\item \code{regional_analysis}: Dataframe with counts and school size metrics for sightings included in the analysis, parsed by regional geostratum.
\item \code{detection_distances}: A dataframe with sighting counts for various
perpendicular distances (in KM) (systematic sightings only).
The \code{percent_beyond} column can be used to identify appropriate truncation distances for this species group.
\item \code{sightings}: Dataframe with all \code{sightings} information for this species group.
}
}
\description{
Summarize sightings for a species (or several)
}
