% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/das_interpolate.R
\name{das_interpolate}
\alias{das_interpolate}
\title{Interpolate DAS data}
\usage{
das_interpolate(das, new_interval = 120, max_ignore = 3600, verbose = FALSE)
}
\arguments{
\item{new_interval}{The interpolation interval, in seconds.}

\item{max_ignore}{The maximum interval (seconds) between two rows of survey data that will be interpolated;
if the time gap exceeds this value, interpolation will not occur for these two rows.
The default is 3600 seconds (1 hour).}

\item{verbose}{Print updates to the Console?}
}
\value{
An interpolated \code{data.frame} of the \code{DAS} data. No formatting has been changed.
}
\description{
This function is not typically called by the user directly; it is called during \code{process_surveys()}.
The user can instruct \code{LTabundR} to interpolate \code{DAS} data in \code{load_survey_settings()}.
}
\details{
param das  A \code{data.frame} of a \code{DAS} survey data file, created by \code{load_das()}.

This function allows you to interpolate the DAS position data
at the onset of processing if your position updates are separated
by large time intervals, which would make spatial effort and stratum assignments less exact.
\code{LTabundR} will interpolate the data using simple-linear methods (i.e., no great-circle calculations), such that
position updates occur every \code{new_interval} seconds or less. If adjacent \code{DAS} rows are from different dates or cruises,
or if the interval between rows exceeds the input \code{max_ignore},
the interpolation routine will skip to the next pair of related rows. Interpolation will only occur for On-Effort rows.
}
