% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group_size_coefficients}
\alias{group_size_coefficients}
\title{Group size coefficients}
\format{
An object of class \code{data.frame} with 162 rows and 21 columns.
}
\usage{
group_size_coefficients
}
\description{
Group size coefficients
}
\details{
A table of group size calibration coefficient values for a variety of species
for a variety of observer codes. This table was provided by Jay Barlow (NOAA NMFS Bioligist), was created
prior to 2016, and has been used for published analyses through 2021.
}
\keyword{datasets}
