% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_settings.R
\name{load_settings}
\alias{load_settings}
\title{Load settings for processing \code{WinCruz} surveys}
\usage{
load_settings(strata = NULL, survey = NULL, cohorts = NULL)
}
\arguments{
\item{strata}{A named list in which each slot is a \code{data.frame} of coordinates for a geostratum polygon.
Each \code{data.frame} must have \code{Lon} and \code{Lat} as the first two columns (in that order),
providing coordinates in decimal degrees in which West and South coordinates are negative.
Other columns are allowed, but the first two need to be \code{Lon} and \code{Lat}.
The name of the slot holding the \code{data.frame} will be used as a reference name for the stratum.
For an example of formatting, see \code{data(strata_cnp)}. If you are working with a standard NOAA survey region,
such as the Central North Pacific (CNP), Eastern Tropical Pacific (ETP), or California Current System (CCS),
you can use built-in polygons available in \code{data(strata_cnp)}, \code{data(strata_etp)}, or \code{data(strata_ccs)}, respectively.
To explore and/or select strata contained within those built-in datasets, use the functions
\code{strata_explore()} and \code{strata_select()}.}

\item{survey}{Survey-wide settings, provided as the list that is generated with the
command \code{load_survey_settings()}. See the documentation for that function for full details.
if nothing is supplied, the default survey-wide settings (see \code{load_survey_settings()})
will be applied.}

\item{cohorts}{Cohort-specific settings, provided as a list of lists.
Each slot contains the settings for a single cohort, which can be generated with the
command \code{load_cohort_settings()}. See the documentation for that function for full details.
Cohort-specific settings apply only to a group of species.
Since you can add as many cohorts to a \code{settings} object as you need,
this allows you to stage your entire analysis and run your code once,
without modifying code between the analysis of each cohort.
If nothing is supplied, the default cohort settings (see \code{load_cohort_settings()})
will be applied to all species.}
}
\value{
A list with three named slots, equivalent to your three input arguments.
Save this output to an object, e.g., "\code{settings}", and pass it to \code{process_surveys()}.
}
\description{
You need the output of this function to pass to \code{process_surveys()},
the primary function for processing \code{WinCruz} data.
}
