% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey_settings.R
\name{load_survey_settings}
\alias{load_survey_settings}
\title{Load survey settings}
\usage{
load_survey_settings(
  out_handling = "remove",
  max_row_interval = 900,
  segment_method = "day",
  segment_target_km = 150,
  segment_max_interval = 48,
  segment_remainder_handling = "segment",
  ship_list = NULL,
  species_codes = NULL,
  group_size_coefficients = NULL,
  smear_angles = FALSE
)
}
\arguments{
\item{out_handling}{This argument allows you to specify
how data occurring outside of geo-strata should be handled.
If this is set to \code{"remove"}, those rows will be filtered out of the data early in the process.
This reduces memory usage, speeds up processing,
and gives you geographic control of how effort and sightings will be summarized.
If this is set to \code{"stratum"}, those data will be assigned to a fake geo-stratum, named \code{"out"}.
Effort in the \code{"out"} stratum will not be segmentized,
but \code{"out"} sightings will be processed and retained in the final datasets.
This setting might be useful if you want to use \code{"out"} data
for survey summaries and/or detection function estimation.
The default is \code{"remove"}, since that saves the most time and memory.
If no geostratum is provided, this setting will be ignored and all rows of data
will be assigned to a stratum called \code{"none"}.}

\item{max_row_interval}{The maximum alloweable time interval, in seconds, between rows before
the function assumes that there has been a break in survey data logging. The default is 900 seconds,
or 15 minutes.}

\item{segment_method}{The two method options are \code{"day"} --
all effort within the same Cruise-StudyArea-Stratum-Year-Effort scenario (i.e., an effort bloc)
will be binned into segments by calendar date -- and \code{"equallength"} --
effort within each unique effort bloc
will be divided into segments of approximately equal length.}

\item{segment_target_km}{If segmentizing by \code{"equallength"},
this field allows you to specify what that target length is, in km.
If segmentizinng by \code{"day"}, this argument is ignored.}

\item{segment_max_interval}{If segmentizing by \code{"equallength"},
this setting allows you to specify the time gaps in effort
that are allowed to be contained within a single segment.
For example, if your goal is a few large segments of equal length
(e.g., 150-km segments, for bootstrap estimation of density variance),
you are probably willing for discrete periods of effort to be concatenated into a single segment,
even if the gaps between effort are as large as 1 or 2 days,
in which case you would set \code{segment_max_interval} to 24 or 48 (hours), respectively.
However, if your goal is many smaller segments (e.g., 5-km segments, for habitat modeling),
you want to ensure that effort is contiguous so that segment locations
can be accurately related to environmental variables,
in which case you would set \code{segment_max_interval} to be very small (e.g., 0.2 hours, or 12 minutes).
Setting this interval to a small number, such as 0.2, also allows
the segmentizing function to overlook momentary breaks in effort,
such as when an unofficial observer logs a sighting.
If segmentizinng by \code{"day"}, this argument is ignored.}

\item{segment_remainder_handling}{If segmentizing by \code{"equallength"},
periods of effectively-contiguous effort (as specified by \code{segment_max_interval})
are unlikely to be perfectly divisible by your \code{segment_target_km};
there is going to be a remainder. You can handle this remainder in three ways:
(1) \code{"disperse"} allows the function to adjust \code{segment_target_km} so that
there is in fact no remainder, effectively dispersing the remainder evenly
across all segments within that period of contiguous effort;
(2) \code{"append"} asks the function to append the remainder to a randomly selected segment,
such that most segments are the target length with the exception of one longer one;
or (3) \code{"segment"} asks the function to simply place the remainder in its own segment,
placed randomly within the period of contiguous effort.
This setting also has a second layer of versatility,
because it can accept a one- or two-element character vector.
If a two-element vector is provided (e.g., \code{c("append","segment")}),
the first element will be used in the event that the remainder is less than or equal to
half your \code{segment_target_km}; if the remainder is more than half that target length,
the second element will be used. This feature allows for replication
of the segmentizing methods in Becker et al. (2010).
If segmentizinng by \code{"day"}, this argument is ignored.}

\item{ship_list}{A \code{data.frame} containing a list of ship names.
If not provided the default version, which was current as of the release of \code{ABUND9} in 2020, will be used (\code{data(ships)}).
Supplied \code{data.frames} must match the column naming structure of \code{data(ships)}.}

\item{species_codes}{A \code{data.frame} containing species codes.
This is an optional input, chiefly used to format species names
in the reporting stage of the workflow (\code{lta_report()} especially).
If the user supplies a \code{data.frame} it must match the column naming structure of \code{data(species_codes)}.}

\item{group_size_coefficients}{A \code{data.frame} of calibration factors.
If not provided, group sizes will not be calibrated.
To use the same coefficients that have been in use at SWFSC and PIFSC up to 2021, see \code{data(group_size_coefficients)}.
Supplied \code{data.frame}'s must match the column naming structure of that built-in dataset.}

\item{smear_angles}{If \code{TRUE} (the default is \code{FALSE}), bearing angles to a
group of animals will be "smeared" by adding a uniformly distributed random number between -5 and +5 degrees.
This has not been used in any recent analyses because observers have not been rounding angles as much as they used to,
according to the release notes for \code{ABUND9}.
It was suggested by Buckland as a method for dealing with rounding which is
especially influential when rounding to zero places many sightings at zero perpendicular distance.}
}
\value{
A list with named slots, equivalent to your input arguments.
Save this output to an object, e.g., "\code{survey_settings}", and pass it to \code{load_settings()}.
}
\description{
This function builds a list of survey-wide settings (applying to all segments
and all sightings of all species, regardless of their cohort designation),
which you will pass to \code{load_settings()}.
}
