% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lta_subgroup.R
\name{lta_subgroup}
\alias{lta_subgroup}
\title{Subgroup-based line transect analysis}
\usage{
lta_subgroup(
  df_sits,
  truncation_distance,
  ss,
  density_segments,
  density_das,
  density_sightings,
  Rg0 = NULL,
  cruz10 = NULL,
  g0_spp = NULL,
  g0_truncation = NULL,
  g0_constrain_shape = FALSE,
  g0_jackknife_fraction = 0.1,
  abundance_area = NULL,
  iterations = 5000,
  output_dir = NULL,
  toplot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df_sits}{(Required.) A \code{data.frame} of sightings you want to use to fit the detection function model.
For false killer whales in Bradford et al. (2020), this is a combination of systematic sightings
prior to 2010 and Phase 1 sightings from 2010 onwards (using the PC protocol).
This \code{dataframe} must have a column named \code{PerpDistKM} with detection distances in km.
No filtering will be applied to these sightings within this function,
so make sure you provide the data pre-filtered. Bradford et al. (2020) used a
single detection function for all populations of false killer whale.}

\item{truncation_distance}{(Required.) The truncation distance, in km, to apply during detection function model fitting.}

\item{ss}{(Required.) A numeric vector of subgroup school sizes to use to find its mean and bootstrapped CV.
In Bradford et al. (2020), data come from all Phase 1 and Phase 2 estimates of subgroup sizes from 2010 onwards.
These estimates are the geometric mean of repeat estimates from separate observers.}

\item{density_segments}{(Required.) The survey segments to be used in density/abundance estimation.
For example, Bradford et al. (2020) used 150-km segments to estimate false killer whale
density in the Hawaiian EEZ in 2017 (these data are available in the built-in dataset
\code{"cnp_150km_1986_2020"}). No filtering will be applied to these segments,
so make sure only the segments you wish to use are included and nothing more.
For example, in the case above, make sure you are only providing systematic segments for the
Hawaiian EEZ in 2017.}

\item{density_das}{(Required.) The complete survey data corresponding to the above segments.
These data will be used to determine the proportion of survey effort occurring in each Beaufort
sea state during Relative \code{g(0)} estimation.}

\item{density_sightings}{(Required.) The encounters to use in density/abundance estimation.
In Bradford et al. (2020), these were the Phase 1 detections of false killer whales
within the population-region-year of interest, e.g., Northwest Hawaiian Island population sightings
within the Hawaiian EEZ in 2017. No filtering will be applied to these sightings,
so make sure only the sightings you wish to use are included and nothing more.}

\item{Rg0}{A \code{data.frame} with estimates of Relative \emph{g(0)} and its CV at each Bft state.
If this input is left \code{NULL}, then these estimates will be produced by the function using the subsequent \code{g0_} inputs.
If this input is not supplied and any of the subsequent \code{g0_} inputs are missing, then \emph{g(0)} will be assumed to be 1.0 with CV of 0.0.
If supplied, this \code{data.frame} has three required columns:
\code{bft} (Beaufort sea state, numbers between 0 and 7),
\code{Rg0} (\emph{Rg(0)} estimates for each Beaufort state),
and \code{Rg0_CV} (the CV of the \emph{Rg(0)} estimate in each Beaufort state). Other columns are allowed but will be ignored.}

\item{cruz10}{A processed \code{cruz} object with short segment lengths, ideally 10 km or less (hence the 10 in the input name).
This \code{cruz} object will be used to estimate \code{Rg(0)}, i.e., the relative trackline detection probability.
Consider using the built-in dataset \code{"noaa_10km_1986_2020"}.}

\item{g0_spp}{A character vector of species codes to use to estimate \code{Rg(0)}.
In most cases this will be a single species, e.g., '033' for false killer whales.
Not required if the \code{Rg0} input is supplied.}

\item{g0_truncation}{The truncation distance to use when estimating \code{Rg(0)}.
In Bradford et al. (2020) this is 5.5 km.}

\item{g0_constrain_shape}{Some \emph{Rg(0)} curves will not decline monotonically
due to sample size issues at low Bft (0-2) or high Bft (5-6) states.
To coerce monotonic decline, set this input to \code{TRUE}, and the function will use a shape-constrained GAM
(\code{scam()} from package \code{scam}) instead of a classic \code{mgcv::gam()}.}

\item{g0_jackknife_fraction}{The proportion of data to leave out within each jackknife permutation for
estimating the CV of \emph{g(0)}. The default is 0.1 (i.e., 10\% of the data, yielding 10 jackknife loops), after Barlow (2015).}

\item{abundance_area}{The area in square km of the region of interest. The density
estimate will be scaled by this area.}

\item{iterations}{Number of iterations to use in the various CV bootstrapping procedures
occurring throughout this function, specifically: Effective Strip Half-Width CV estimation,
school size CV estimation, weighted \code{g(0)} CV estimation, encounter rate estimation, and density/abundance estimation.}

\item{output_dir}{The path in which results \code{RData} files should be stored. If left \code{NULL}, no results will be stored.
To use your current working directory, simply provide \code{""}.}

\item{toplot}{A Boolean, with default \code{FALSE}, indicating whether to plot various
aspects of the analysis.}

\item{verbose}{A Boolean, with default \code{TRUE}, indicating whether to print status updates to the Console.}
}
\value{
A list.
\enumerate{
\item \code{D}: The estimate of density.
\item \code{D_CV}: The CV of the density estimate.
\item \code{D_L95}: The lower 95\% confidence interval of density using the BCA method.
\item \code{D_U95}: The upper 95\% confidence interval of density using the BCA method.
\item \code{N}: The estimate of abundance.
\item \code{N_CV}: The CV of the abundance estimate.
\item \code{N_L95}: The lower 95\% confidence interval of abundance using the BCA method.
\item \code{N_U95}: The upper 95\% confidence interval of abundance using the BCA method.
\item \code{ER}: The estimate of the encounter rate.
\item  \code{ESW} = The estimate of the Effective Strip Width (km).
\item  \code{ESW_CV} = Estimate of ESW CV, based on standard deviation of bootstrap estimates of ESW.
\item \code{ss} = Mean subgroup size estimate.
\item \code{n}: Number of sightings used in density estimation.
\item \code{L}: Survey effort, in km, used in density estimation.
\item \code{n_segments}: Number of effort segments used in density estimation.
\item \code{g0}: The empirical weighted mean of \emph{g(0)} for the point estimate, based on sightings conditions in \code{density_segments}.
\item \code{g0_cv}: The CV of this estimate of the point estimate of the weighted mean of \emph{g(0)}, as estimated by an MCMC routine.
\item \code{g0_details}: A \code{list} with detailed results from \code{Rg(0)} estimation (see output details in \code{?g0_model}).
\item \code{df}: A \code{list} with detailed results from detection function estimation (see output details in \code{?df_fit}).
\item \code{bootstraps}: A named \code{list} with the bootstrapped values for \code{esw} (effective strip half-width),
\code{ss} (subgroup size), \code{g0} (relative g(0)), \code{er} (encounter rate), \code{D} (density), and \code{N} (abundance).
\item \code{iterations}: number of bootstrap iterations used for CV estimation.
}
See the \href{https://emk-noaa.github.io/LTAvignette/subgroup-based-analysis.html}{online vignette} for more details
}
\description{
A flexible routine for carrying out subgroup-based line-transect analysis
using the methodologies described for Hawai'ian false killer whales,
\emph{Pseudorca crassidens}, in \href{https://www.fisheries.noaa.gov/inport/item/59592}{Bradford et al. (2020)}.
The function returns an estimate
of density and abundance -- along with estimates of intermediate parameters --
with a CV derived from a bootstrapping routine.
As part of this process, relative trackline detection probability (\code{g(0)}) is modeled
as a function of Beaufort sea state (using \code{LTabndR} function \code{g0_model()}),
then a weighted \code{g(0)} and its CV are estimated using \code{LTabundR} function \code{g0_weighted()}.
}
\details{
This function performs the following operations:
\enumerate{
\item Fits a detection function to \code{df_sits} without covariates, using the \code{LTabundR} function \code{df_fit()},
in order to estimate the effective strip half-width (ESW).
\item Conducts bootstrap re-sampling of the detection function fitting routine in order to estimate the CV of ESW.
\item Estimates the arithmetic mean of subgroup school size based on the \code{ss} input.
\item Creates a bootstrap-resampled distribution of subgroup school sizes, with which CV is estimated.
\item Models the relative g(0) in different survey conditions using the \code{LTabundR} function \code{g0_model()}.
This function also estimates the CV of the Rg(0) estimate in each Beaufort sea state using jackknife resampling.
\item Estimates the encounter rate (subgroup detections / trackline surveyed).
\item Creates a bootstrap-resampled distribution of encounter rate estimates.
\item Calculates a weighted \code{g(0)} estimate according to the proportion of effort occurring in each Beaufort sea state,
then uses an automated parameter optimization routine (see details in \code{LTabundR} function \code{g0_weighted()}) to
estimate the CV of the weighted \code{g(0)} estimate.
\item Creates a bootstrap-resampled distribution of the weighted \code{g(0)} estimate.
\item Estimates density using the best estimates of effective strip half-width, school size, \code{g(0)}, and the encounter rate.
\item Estimates abundance by scaling the density estimate by the provided \code{abundance_area}.
\item Creates a bootstrap-resampled distribution of the density estimate by
iteratively drawing values (without replacement) from the resampled distributions
of the constituent parameters of the density equation.
\item Creates a bootstrap-resampled distribtion of the abundance estimate by
scaling the density distribution by \code{abundance_area}.
}
}
