% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polygon.R
\name{process_polygon}
\alias{process_polygon}
\title{Format a CSV into a polygon for mapping}
\usage{
process_polygon(
  polygon_dataframe,
  remove_land = TRUE,
  coerce_lons = FALSE,
  toplot = FALSE
)
}
\arguments{
\item{polygon_dataframe}{A dataframe in which the first two columns are \code{Lon} and \code{Lat},
providing coordinates in decimal degrees in which South and West coordinates are negative.
Other columns are allowed, but the first two need to be \code{Lon} and \code{Lat} (in that order).}

\item{remove_land}{If \code{TRUE}, the area of land occurring within the polygon will
be subtracted from the polygon's area, so that the reported area represents aquatic habitat only.}

\item{coerce_lons}{If \code{TRUE} (not the default), all longitudes will be coerced to negative values, e.g., -185.
This can avoid complications when strata or survey data span the international date line.}

\item{toplot}{If \code{TRUE} (not the default), a map of the polygon will be plotted,
and -- if \code{remove_land} is \code{TRUE} -- any land inside will be plotted in grey.}
}
\value{
A list with four named slots:
\enumerate{
\item \code{coords} The input dataframe.
\item \code{polygon}  The polygon created by \code{sf::st_polygon()}
\item \code{sf}  The \code{sf} version of the polygon that will be passed to mapping functions.
\item \code{km2}  The area of the polygon, in square kilometers.
}
}
\description{
This is an internal function typically not called by a user directly.
It takes a dataframe of polygon coordinates and returns a list with the polygon formatted in various ways,
as well as a calculation of the area of the polygon \emph{with land area removed} (optionally).
}
