% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sightings.R
\name{process_sightings}
\alias{process_sightings}
\title{Process survey sightings}
\usage{
process_sightings(
  cruz,
  calibrate = NULL,
  calibrate_floor = NULL,
  geometric_mean = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cruz}{A \code{cruz} object passed from \code{segmentize()}.}

\item{calibrate}{An argument allowing you to override the settings contained within the \code{cruz} object.
This argument accepts a Boolean; if \code{TRUE}, school size calibration will be attempted,
but only if calibration coefficients are provided in \code{cruz$settings$survey}.}

\item{calibrate_floor}{Another argument allowing for settings override.
This argument accepts a number indicating the minimum raw school size estimate
for which school size calibration will be attempted.}

\item{geometric_mean}{Another argument allowing settings override.
This argument accepts a Boolean; if \code{TRUE}, geometric means will be calculated instead of arithmetic means.
If school size calibration is carried out, the geometric mean will be weighted by calibration variance, such that
estimates from observers with low variance will receive more weight.}

\item{verbose}{Boolean, with default \code{FALSE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A modified \code{cruz} object, in which a \code{sightings} dataframe
(one row for each species in each sighting, with school size estimates averaged and calibrated according to cohort-specific \code{settings})
has been added to each \verb{<cohort>} list. The \code{sightings} dataframe has a new column, \code{included}, indicating whether or not
a sighting will be included in an analysis according to the inclusion criteria specified in cohort-specific \code{settings},
as well as a new column, \code{ss_valid}, indicating whether or not the school size estimate
for this sighting is valid and appropriate for use in abundance estimation and/or detection function fitting
with a school-size covariate. All columns are described in detail within the
documentation for the \code{LTabundR} function, \code{process_surveys()}.
}
\description{
This is an internal function typically not called by a user directly.
It is the fifth subroutine called within \code{process_surveys()}, after \code{segmentize()} and before \code{process_subgroups()}.
}
