% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_subgroups.R
\name{process_subgroups}
\alias{process_subgroups}
\title{Analyze subgroup detections for false killer whales}
\usage{
process_subgroups(cruz, phase_edits = NULL, verbose = TRUE)
}
\arguments{
\item{cruz}{A \code{cruz} object passed from \code{process_surveys()}.}

\item{phase_edits}{An optional input: a \code{data.frame} indicating manual adjustments to the protocol phase that data are assigned to.
This \code{data.frame} is produced by the \code{LTabundR} function \code{subgroup_phases()}; see its documentation for formatting.
If this input is supplied, the \code{process_subgroups()} function will first automatically assign data to phases,
then redact those assignments using this input. The recommended workflow here is:
(1) process your survey data without supplying \code{phase_edits};
(2) review the automatic phase assignments using the \code{subgroup_phases()} function;
(3) if revisions are needed, stage those edits using the same function.
(4) re-run \code{process_subgroups()}, this time supplying the \code{phase_edits} input.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A finalized \code{cruz} object.
If subgroups are found, a \code{subgroups} slot is added to the analysis list for a cohort.
This \code{subgroups} slot holds a list with three dataframes:
\enumerate{
\item \code{events}, in which each row is a school size estimate for a single subgroup during a single phase -- 1 or 2 -- within a single sighting;
\item \code{subgroups}, in which each row is a single phase for a single subgroup, with all school size estimates averaged together (both arithmetically and geometrically);
\item \code{sightings}, in which each row is a school size estimate for a single phase for a single sighting, with all subgroup school sizes summed together.
}
Phase is determined simply according to the \code{OnEffort} column. If it is \code{TRUE}, \code{Phase} is 1; if not, \code{Phase} is 2.
}
\description{
Subgroups are found based on \code{DAS} event code "\code{G}" and associated events.
Subgroup sizes are estimated for each phase (based upon the false killer whale protocol),
and both arithmetic and geometric means are provided.
\cr \cr
This is an internal function typically not called by a user directly.
It is the final processing subroutine called within \code{process_surveys()}, after the subroutine \code{process_sightings()}.
}
