% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_subgroups.R
\name{process_subgroups}
\alias{process_subgroups}
\title{Analyze subgroup detections for false killer whales}
\usage{
process_subgroups(cruz, verbose = TRUE)
}
\arguments{
\item{cruz}{A \code{cruz} object passed from \code{process_surveys()}.}

\item{verbose}{Boolean, with default \code{TRUE}, indicating whether or not updates should be printed to the Console.}
}
\value{
A finalized \code{cruz} object.
If subgroups are found, a \code{subgroups} slot is added to the analysis list for a cohort.
This \code{subgroups} slot holds a list with three dataframes:
\enumerate{
\item \code{events}, in which each row is a school size estimate for a single subgroup during a single phase -- 1 or 2 -- within a single sighting;
\item \code{subgroups}, in which each row is a single phase for a single subgroup, with all school size estimates averaged together (both arithmetically and geometrically);
\item \code{sightings}, in which each row is a school size estimate for a single phase for a single sighting, with all subgroup school sizes summed together.
}
}
\description{
Subgroups are found based on \code{DAS} event code "\code{G}" and associated events.
Subgroup sizes are estimated for each phase (based upon the false killer whale protocol),
and both arithmetic and geometric means are provided.
\cr \cr
This is an internal function typically not called by a user directly.
It is the final processing subroutine called within \code{process_surveys()}, after the subroutine \code{process_sightings()}.
}
