% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_segments.R
\name{strata_segments}
\alias{strata_segments}
\title{Display/map stratum assignments for segments}
\usage{
strata_segments(cruz, cohort = 1)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to display, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}
}
\value{
A \code{ggplot2} object.
}
\description{
Display/map stratum assignments for segments
}
