% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_bft.R
\name{summarize_bft}
\alias{summarize_bft}
\title{Summarize \code{Wincruz} by Beaufort Sea State}
\usage{
summarize_bft(cruz, cohort = 1)
}
\arguments{
\item{cruz}{Your \code{cruz} object (produced from \code{LTabundR::process_surveys()}).}

\item{cohort}{The cohort whose data you would like to summarize, provided as a number indicating which slot in \code{cruz$cohorts} should be referenced.}
}
\value{
A list with various summary tables:
\enumerate{
\item \code{overall}: Overall summary of effort, parsed by Beaufort sea state, for the entire dataset.
Three columns: \code{bftr} (rounded Beaufort), \code{km}, \code{prop} (the fraction of effort for this sea state).
\item \code{by_year}: Same table as \code{overall}, this time parsed by \code{year}.
\item \code{by_cruise}: Same table as \code{overall}, this time parsed by geostratum.
\item \code{details}: Same table as \code{overall}, this time parsed by each unique \code{Cruise}-\code{year}-\code{stratum} combination.
}
}
\description{
Calulcate the distance and proportion of effort within each Beaufort Sea State
within your \code{Wincruz} survey data.
}
